% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyWindows.R
\name{applyWindows}
\alias{applyWindows}
\title{apply a function to several windows of a data.frame}
\usage{
applyWindows(
  ds,
  FUN,
  prevRes = list(),
  winInfo = get_winInfo(nrow(ds)),
  isVerbose = TRUE,
  ...
)
}
\arguments{
\item{ds}{data.frame to iterate}

\item{FUN}{function to apply to subsets of the data.frame taking a subset of
the data.frame as first argument the second: a one-row data.frame with window
information (iWindow, dayStart, dayEnd, iRecStart, iRecEnd, iCentralRec) the
third: most recent valid result of calls FUN. Valid is a non-NULL result.}

\item{prevRes}{initial values for the list that
is carried between windows}

\item{isVerbose}{set to FALSE to suppress messages}

\item{...}{further arguments to FUN}
}
\description{
apply a function to several windows of a data.frame
}
