% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitioningLasslop10.R
\name{.computeRecoNight}
\alias{.computeRecoNight}
\title{recompute ecosystem respiration at nighttime with nighttime reference}
\usage{
.computeRecoNight(dsAnsFluxes, dsR, dsTempSens)
}
\arguments{
\item{dsAnsFluxes}{data.frame with columns Reco and sdReco}

\item{dsR}{data.frame with columns isNight, and Temp}

\item{dsTempSens}{data.frame with column \code{iCentralRec}, \code{E0}, and \code{RRef}}
}
\value{
data.frame with updated night-time \code{Reco} and \code{sdReco}
}
\description{
recompute ecosystem respiration at nighttime with nighttime reference
}
\details{
If \code{partGLControl$useNightimeBasalRespiration == TRUE} then
nighttime respiration is recomputed based on basal respiration inferred
from nighttime data.
}
