% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitioningLasslop10.R
\name{.partGPAssociateSpecialRows}
\alias{.partGPAssociateSpecialRows}
\title{associate each row with the previous and next row from a subset of rows}
\usage{
.partGPAssociateSpecialRows(iRowsSpecial, nRec)
}
\arguments{
\item{iRowsSpecial}{ordered unique integer vector specifying the rows for which
some special data is available}

\item{nRec}{integer scalar of number of rows in the full data.frame}
}
\value{
a dataframe with index of previous and next rows inside the subset
}
\description{
associate each row with the previous and next row from a subset of rows
}
