\name{fAddNCFVar}
\alias{fAddNCFVar}
\title{Add variable from NetCDF file to data frame}

\usage{fAddNCFVar(Data.F, Var.s, FileName.s, Dir.s)}
\arguments{
  \item{Data.F}{Data frame}
  \item{Var.s}{Variable name}
  \item{FileName.s}{NetCDF file name}
  \item{Dir.s}{Directory}
}

\value{Data frame with new nc variable added.}

\author{AMM, KS
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





