\name{fCalcAVPfromVMFandPress}
\alias{fCalcAVPfromVMFandPress}
\title{fCalcAVPfromVMFandPress}
\description{Calculate AVP from VMF and Press}
\usage{fCalcAVPfromVMFandPress(VMF.V.n, Press.V.n)}
\arguments{
  \item{VMF.V.n}{Vapor mole fraction (VMF, mol/mol)}
  \item{Press.V.n}{Atmospheric pressure (Press, hPa)}
}

\value{Data vector of actual vapor pressure (AVP, hPa (mbar))}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





