\name{fCalcETfromLE}
\alias{fCalcETfromLE}
\title{fCalcETfromLE}
\description{Calculate ET from LE and Tair}
\usage{fCalcETfromLE(LE.V.n, Tair.V.n)}\arguments{
  \item{LE.V.n}{Data vector of latent heat (LE, W m-2)}
  \item{Tair.V.n}{Data vector of air temperature (Tair, degC)}
}

\value{Data vector of evapotranspiration (ET, mmol H20 m-2 s-1)}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





