\name{fCalcExtRadiation}
\alias{fCalcExtRadiation}
\title{fCalcExtRadiation}
\description{Calculate the extraterrestrial solar radiation with the eccentricity correction}
\usage{fCalcExtRadiation(DoY.V.n)}
\arguments{
  \item{DoY.V.n}{Data vector with day of year (DoY)}
}

\value{Data vector of extraterrestrial radiation (ExtRad, W_m-2)}

\author{AMM
Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut Šigut <sigut.l@czechglobe.cz> [ctb]}





