\name{fCalcPotRadiation}
\alias{fCalcPotRadiation}
\title{Calculate the potential radiation }

\usage{fCalcPotRadiation(DoY.V.n, Hour.V.n, Lat_deg.n, Long_deg.n, TimeZone_h.n)}
\arguments{
  \item{DoY.V.n}{Data vector with day of year (DoY)}
  \item{Hour.V.n}{Data vector with time as decimal hour}
  \item{Lat_deg.n}{Latitude in (decimal) degrees}
  \item{Long_deg.n}{Longitude in (decimal) degrees}
  \item{TimeZone_h.n}{Time zone (in hours)}
}

\value{Data vector of potential radiation (PotRad, W_m-2)}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





