\name{fCalcRHfromAVPandTair}
\alias{fCalcRHfromAVPandTair}
\title{fCalcRHfromAVPandTair}
\description{Calculate relative humidity from actual vapour pressure and air tempature}
\usage{fCalcRHfromAVPandTair(AVP.V.n, Tair.V.n)}
\arguments{
  \item{AVP.V.n}{Data vector of actual vapour pressure (AVP, hPa (mbar))}
  \item{Tair.V.n}{Data vector of air temperature (Tair, degC)}
}

\value{Data vector of relative humidity (rH, \%)}

\author{AMM
Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





