\name{fCalcVPDfromRHandTair}
\alias{fCalcVPDfromRHandTair}
\title{fCalcVPDfromRHandTair}
\description{Calculate VPD from rH and Tair}
\usage{fCalcVPDfromRHandTair(RH.V.n, Tair.V.n)}
\arguments{
  \item{RH.V.n}{Data vector of relative humidity (rH, \%)}
  \item{Tair.V.n}{Data vector of air temperature (Tair, degC)}
}

\value{Data vector of vapour pressure deficit (VPD, hPa (mbar))}

\author{AMM
Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut Šigut <sigut.l@czechglobe.cz> [ctb]}





