\name{fCheckColNames}
\alias{fCheckColNames}
\title{Check if specified columns exists in data frame}

\usage{
fCheckColNames(Data.F,
ColNames.V.s,
CallFunction.s = "")
}
\arguments{
  \item{Data.F}{Data frame}
  \item{ColNames.V.s}{Vector of variables to be checked}
  \item{CallFunction.s}{Name of function called from}
}

\value{Function stops on errors.}

\author{AMM
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





