\name{fCheckColNum}
\alias{fCheckColNum}
\title{fCheckColNum}
\description{Check if specified columns are numeric}
\usage{fCheckColNum(Data.F, ColNames.V.s, CallFunction.s = "")}
\arguments{
  \item{Data.F}{Data frame}
  \item{ColNames.V.s}{Vector of variables to be checked, with 'none' as dummy}
  \item{CallFunction.s}{Name of function called from}
}

\value{Function stops on errors.}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





