\name{fCheckColPlausibility}
\alias{fCheckColPlausibility}
\title{Check plausibility of common (eddy) variables}

\usage{
fCheckColPlausibility(Data.F,
VarName.V.s,
CallFunction.s = "")
}
\arguments{
  \item{Data.F}{Data frame}
  \item{VarName.V.s}{Variable (column) names}
  \item{CallFunction.s}{Name of function called from}
}
\details{Variables CONTAINing the following abbreviations are checked for plausibility}
\value{Function produces warnings if variable outside range.}

\author{AMM
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





