\name{fCheckHHTimeSeries}
\alias{fCheckHHTimeSeries}
\title{Check half-hourly time series data}

\usage{
fCheckHHTimeSeries(Time.V.p,
CallFunction.s = "")
}
\arguments{
  \item{Time.V.p}{Time vector in POSIX format}
  \item{CallFunction.s}{Name of function called from}
}
\details{The time stamp needs to be provided in POSIX time format,}
\value{Function stops on errors.}

\author{AMM
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





