\name{fCheckValNum}
\alias{fCheckValNum}
\title{fCheckValNum}
\description{Check if variable is a numeric}
\usage{fCheckValNum(Value.n)}\arguments{
  \item{Value.n}{Value to be checked if numeric (but can also be NA of any type)}
}
\details{See test_CheckValue.R for more details.}
\value{Boolean value if true of false.}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





