\name{fCheckValString}
\alias{fCheckValString}
\title{Check if variable is a non-empty character string}

\usage{
fCheckValString(Value.s)
}
\arguments{
  \item{Value.s}{Value to be checked if string}
}
\details{See test_CheckValue.R for more details.}
\value{Boolean value if true of false.}

\author{AMM
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





