\name{fConvertCtoK}
\alias{fConvertCtoK}
\title{fConvertCtoK}
\description{Convert degree Celsius to degree Kelvin}
\usage{fConvertCtoK(Celsius.V.n)}
\arguments{
  \item{Celsius.V.n}{Data vector in Celsius (degC)}
}

\value{Data vector in temperature Kelvin (Temp_K, degK)}

\author{AMM
Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





