\name{fConvertGapsToNA}
\alias{fConvertGapsToNA}
\title{Converts all gap flags to NA}

\usage{fConvertGapsToNA(Data.F, GapFlag.n = -9999)}
\arguments{
  \item{Data.F}{Data frame to be converted}
  \item{GapFlag.n}{Flag value used for gaps, defaults to -9999.0}
}

\value{Data frame with NAs instead of gaps.}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





