\name{fConvertGlobalToVisible}
\alias{fConvertGlobalToVisible}
\title{fConvertGlobalToVisible}
\description{Partition global (solar) radiation into only visible (the rest is UV and infrared)}
\usage{fConvertGlobalToVisible(Global.V.n)}
\arguments{
  \item{Global.V.n}{Data vector of global radiation (W m-2)}
}

\value{Data vector of visible part of solar radiation (VisRad, W m-2)}

\author{AMM
Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut Šigut <sigut.l@czechglobe.cz> [ctb]}





