\name{fConvertKtoC}
\alias{fConvertKtoC}
\title{Convert degree Kelvin to degree Celsius}

\usage{fConvertKtoC(Kelvin.V.n)}
\arguments{
  \item{Kelvin.V.n}{Data vector in Kelvin (°K)}
}

\value{Data vector in temperature Celsius (Temp_C, °C)}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





