\name{fConvertNAsToGap}
\alias{fConvertNAsToGap}
\title{Converts all NAs to gap flag}

\usage{fConvertNAsToGap(Data.F, GapFlag.n = -9999)}
\arguments{
  \item{Data.F}{Data frame to be converted}
  \item{GapFlag.n}{Flag value used for gaps, defaults to -9999.0}
}

\value{Data frame with gap value instead of NAs.}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





