\name{fConvertVisibleWm2toPhotons}
\alias{fConvertVisibleWm2toPhotons}
\title{fConvertVisibleWm2toPhotons}
\description{Convert units of visible radiation from irradiance to photons flux}
\usage{fConvertVisibleWm2toPhotons(Wm2.V.n)}
\arguments{
  \item{Wm2.V.n}{Data vector in units of irradiance (W m-2)}
}

\value{Data vector in units of photons flux (PPFD, umol photons m-2 s-1)}

\author{Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





