\name{fFilterAttr}
\alias{fFilterAttr}
\title{fFilterAttr}
\description{filter, i.e. subset rows, a data.frame with keeping attributes (e.g. units) of the columns}
\usage{fFilterAttr(x, isFiltered.v)}
\arguments{
  \item{x}{data frame to filter}
  \item{isFiltered.v}{boolean vector specifying which rows to keep}
}

\value{\code{x[,isFiltered.v]} with column attributes preserved}

\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





