% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{fillNAForward}
\alias{fillNAForward}
\title{replace NA by value of previous record}
\usage{
fillNAForward(x, firstValue = median(x, na.rm = TRUE))
}
\arguments{
\item{x}{numeric vector to fill NAs}

\item{firstValue}{value to be used for NA at the beginning of x}
}
\value{
numeric vector with NAs replaced
}
\description{
replace NA by value of previous record
}
