\name{fInterpolateGaps}
\alias{fInterpolateGaps}
\title{Interpolate linearly between gaps, with constant values at beginning/end}

\usage{fInterpolateGaps(Data.V.n)}
\arguments{
  \item{Data.V.n}{Numeric vector with gaps (missing values, NAs)}
}

\value{Numeric with NAs linearly interpolated.}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





