% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileHandlingFormats.R
\name{fLoadAmeriflux22}
\alias{fLoadAmeriflux22}
\title{Read basic variables from Ameriflux standard (as of 2022) files}
\usage{
fLoadAmeriflux22(file_path, ...)
}
\arguments{
\item{file_path}{scalar string: the path to the csv file}

\item{...}{further arguments to \code{\link{read_csv}}}
}
\value{
Data.Frame with columns 
  DateTime, NEE,	Rg,	Tair,	rH,	VPD, Ustar, LE, H
}
\description{
NEE is read from FC, Rg from SW_in, VPD is computed from RH and Tair.
Non-storage corrected LE and H are read.
}
