\name{fSetFile}
\alias{fSetFile}
\title{Set file name with path and check if directory and/or file exists}

\usage{
fSetFile(FileName.s,
Dir.s,
IO.b,
CallFunction.s = "")
}
\arguments{
  \item{FileName.s}{File name}
  \item{Dir.s}{Directory}
  \item{IO.b}{Input/output flag, TRUE for input, FALSE for output}
  \item{CallFunction.s}{Name of function called from}
}

\value{Returns name of file with complete path.}

\author{AMM
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





