\name{fSplitDateTime}
\alias{fSplitDateTime}
\title{fSplitDateTime}
\description{Replace Column DateTime by columns Year, DoY, and Hour in a data.frame}
\usage{fSplitDateTime(df)}
\arguments{
  \item{df}{data.frame with column DateTime}
}
\details{This function helps exporting to the format required by the
REddyProc web interface with columns Year, DoY, and Hour}


\author{Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]}



\seealso{\code{\link{fWriteDataframeToFile}}}

