\name{getAnnualSeasonUStarMappingFromDistributionResult}
\alias{getAnnualSeasonUStarMappingFromDistributionResult}
\title{getAnnualSeasonUStarMappingFromDistributionResult}
\description{extract mapping season -> uStar columns from Distribution result (\code{\link{sEstUstarThresholdDistribution}}}
\usage{getAnnualSeasonUStarMappingFromDistributionResult(uStarTh)}\arguments{
  \item{uStarTh}{
}
}
\details{rows with no threshold get the aggregated value, i.e. the median of columns of other years}


\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





