\name{getFilledExampleDETha98Data}
\alias{getFilledExampleDETha98Data}
\title{getFilledExampleDETha98Data}
\description{get the gapfilled version of the Example_DETha98 example data and create if if not existing yet}
\usage{getFilledExampleDETha98Data()}

\value{example data.frame Example_DETha98 processed by gapfilling. }

\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





