\name{getYearOfSeason}
\alias{getYearOfSeason}
\title{getYearOfSeason}

\usage{getYearOfSeason(seasonFactor.v, sDateTime.v)}
\arguments{
  \item{seasonFactor.v}{factor vector of length data: for each record which season it belongs to}
  \item{sDateTime.v}{POSIX.t vector of length data: for each record: center of half-hour period}
}

\value{named integer vector, with names corresponding to seasons}

\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





