% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partGLFit.R
\name{isValidNightRecord}
\alias{isValidNightRecord}
\title{compute logical vector of each rows in ds is its a valid night record}
\usage{
isValidNightRecord(ds)
}
\arguments{
\item{ds}{data.frame with columns isNight, NEE, Temp (degC)}
}
\value{
logical vector of length nrow(ds)
}
\description{
compute logical vector of each rows in ds is its a valid night record
}
\details{
For robustness, data is trimmed to conditions at temperature > 1 degC
but only timmed if there are more at least 12 records left
}
