\name{LightResponseCurve_getParameterNames}
\alias{LightResponseCurve_getParameterNames}
\title{LightResponseCurve getParameterNames}
\description{return the parameter names used by this Light Response Curve Funciton}
\usage{LightResponseCurve_getParameterNames()}

\value{string vector of parameter names. Positions are important.}

\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





