% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_base.R
\docType{class}
\name{LightResponseCurveFitter-class}
\alias{LightResponseCurveFitter-class}
\alias{LightResponseCurveFitter}
\title{R5 reference parent class for describing the NEP ~ PAR relationship}
\description{
R5 reference parent class for describing the NEP ~ PAR relationship
}

