\name{LightResponseCurveFitter}
\alias{LightResponseCurveFitter}
\title{LightResponseCurveFitter}

\description{
Constructs an instance of class \code{\link{LightResponseCurveFitter-class}}.
However, better construct specialized descendants.
}

\usage{LightResponseCurveFitter(...)}
\arguments{
  \item{\dots}{not used
}
}

\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





