\name{LightResponseCurveFitter_getParameterNames}
\alias{LightResponseCurveFitter_getParameterNames}
\title{LightResponseCurveFitter getParameterNames}
\description{return the parameter names used by this Light Response Curve Funciton}
\usage{LightResponseCurveFitter_getParameterNames()}

\value{string vector of parameter names. Positions are important.
\item{k}{VPD effect}
\item{beta}{saturation of GPP at high radiation}
\item{alpha}{initial slope}
\item{RRef}{basal respiration (units of provided NEE,
usually mumol CO2 m-^-2 s^-2)}
\item{E0}{temperature sensitivity estimated
from night-time data (K)
}}

\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





