\name{LightResponseCurveFitter_getPriorLocation}
\alias{LightResponseCurveFitter_getPriorLocation}
\title{LightResponseCurveFitter getPriorLocation}
\description{return the prior distribution of parameters}
\usage{LightResponseCurveFitter_getPriorLocation(NEEDay, 
    RRefNight, E0)}
\arguments{
  \item{NEEDay}{numeric vector of daytime NEE}
  \item{RRefNight}{numeric scalar of basal respiration estimated from night-time data}
  \item{E0}{numeric scalar of night-time estimate of temperature sensitivity}
}

\value{a numeric vector with prior estimates of the parameters}

\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





