% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_base_optim.R
\name{LRC_computeCost}
\alias{LRC_computeCost}
\title{Computing residual sum of squares for predictions vs. data of NEE}
\usage{
LRC_computeCost(
  thetaOpt,
  theta,
  iOpt,
  flux,
  sdFlux,
  parameterPrior,
  sdParameterPrior,
  ...
)
}
\arguments{
\item{thetaOpt}{numeric vector of optimized parameters}

\item{theta}{numeric vector of parameters}

\item{iOpt}{integer vector of positions of parameters that are optimized}

\item{flux}{numeric: NEP (-NEE) or GPP time series \link{umolCO2 / m2 / s},
should not contain NA}

\item{sdFlux}{numeric: standard deviation of Flux \link{umolCO2 / m2 / s},
should not contain NA}

\item{parameterPrior}{numeric vector along theta: prior estimate of
parameter (range of values)}

\item{sdParameterPrior}{standard deviation of parameterPrior}

\item{...}{other arguments to
\code{\link[=LRC_predictLRC]{LRC_predictLRC()}}, such as VPD0, fixVPD}

\item{weightMisfitPar2000}{weight of misfit of difference between
saturation and prediction at PAR = 2000}
}
\value{
numeric: residual sum of squares
}
\description{
Computing residual sum of squares for predictions vs. data of NEE
}
