% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_base.R
\name{LRC_fitLRC}
\alias{LRC_fitLRC}
\title{Optimize rectangular hyperbolic light response curve in one window}
\usage{
LRC_fitLRC(
  dsDay,
  E0,
  sdE0,
  RRefNight,
  controlGLPart = partGLControl(),
  lastGoodParameters = rep(NA_real_, 7L)
)
}
\arguments{
\item{dsDay}{data.frame with columns NEE, Rg, Temp_C, VPD, and no NAs in NEE}

\item{E0}{temperature sensitivity of respiration}

\item{sdE0}{standard deviation of E_0.n}

\item{RRefNight}{basal respiration estimated from night time data}

\item{controlGLPart}{further default parameters (see \code{\link[=partGLControl]{partGLControl()}})}

\item{lastGoodParameters}{numeric vector returned by last reasonable fit}
}
\value{
A list with the following components:
\itemize{
\item \code{thetaOpt}          : vector of optimized parameters, including the fixed ones and E0
\item \code{iOpt}              : vector of positions of parameters that are optimized,
including E0, which has been optimized prior to this function
\item \code{thetaInitialGuess} : initial guess from data
\item \code{covParms}          : matrix of the covariance matrix of parameters,
including E0
\item convergence         : code specifying convergence problems
}
\itemize{
\item 0      : good convergence
\item 1-1000 : see \code{\link[=optim]{optim()}}
\item 1001   : too few bootstraps converged
\item 1002   : fitted parameters were outside reasonable bounds
\item 1003   : too few valid records in window
\item 1004   : near zero covariance in bootstrap indicating bad fit
\item 1005   : covariance from curvature of fit yielded negative variances
indicating bad fit
\item 1006   : prediction of highest PAR in window was far from saturation
indicating insufficient data to constrain LRC
\item 1010   : no temperature-respiration relationship found
\item 1011   : too few valid records in window (from different location:
partGLFitLRCOneWindow)
}
}
\description{
Optimize rectangular hyperbolic light response curve in one window
}
\details{
Optimization is performed for three initial parameter sets that
differ by beta0 (* 1.3, * 0.8). From those three, the optimization result is
selected that yielded the lowest misfit.

Starting values are:
\itemize{
\item \code{k} = 0,
\item \code{beta} = interpercentileRange(0.03, 0.97) of respiration,
\item \code{alpha} = 0.1,
\item \code{R_ref} from nightTime estimate.
\item \code{E0} is fixed to the night-time estimate, but varies for estimating parameter uncertainty.
}
}
\seealso{
\code{\link[=partGLFitLRCWindows]{partGLFitLRCWindows()}}, \code{\link[=LRC_optimLRCBounds]{LRC_optimLRCBounds()}}
}
