% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_base.R
\name{LRC_getOptimizedParameterPositions}
\alias{LRC_getOptimizedParameterPositions}
\title{Get the positions of the parameters to optimize for given Fixed}
\usage{
LRC_getOptimizedParameterPositions(isUsingFixedVPD, isUsingFixedAlpha)
}
\arguments{
\item{isUsingFixedVPD}{boolean scalar: if TRUE, VPD effect set to zero and is not optimized}

\item{isUsingFixedAlpha}{boolean scalar: if TRUE, initial slope is fixed and is not optimized}
}
\value{
integer vector of positions in parameter vector
}
\description{
If subclasses extend the parameter vector, they need to override this method.
}
