% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_base.R
\name{LRC_getParamNames}
\alias{LRC_getParamNames}
\title{LRC_getParamNames}
\usage{
LRC_getParamNames()
}
\description{
LRC_getParamNames
}
\details{
\itemize{
\item \code{k}     : VPD effect
\item \code{beta}  : saturation of GPP at high radiation
\item \code{alpha} : initial slope
\item \code{RRef}  : basal respiration (units of provided NEE, usually mumol CO2 m-2 s-2)
\item \code{E0}    : temperature sensitivity estimated from night-time data (K)
}
}
