% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_base.R
\name{LRC_getPriorLocation}
\alias{LRC_getPriorLocation}
\title{LRC_getPriorLocation}
\usage{
LRC_getPriorLocation(NEEDay, RRefNight, E0)
}
\arguments{
\item{NEEDay}{numeric vector of daytime NEE}

\item{RRefNight}{numeric scalar of basal respiration estimated from night-time data}

\item{E0}{numeric scalar of night-time estimate of temperature sensitivity}
}
\value{
a numeric vector with prior estimates of the parameters
}
\description{
LRC_getPriorLocation
}
