% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_base.R
\name{LRC_getPriorScale}
\alias{LRC_getPriorScale}
\title{LRC_getPriorScale}
\usage{
LRC_getPriorScale(thetaPrior, medianRelFluxUncertainty, nRec, ctrl)
}
\arguments{
\item{thetaPrior}{numeric vector of location of priors}

\item{medianRelFluxUncertainty}{numeric scalar: median across the
relative uncertainty of the flux values, i.e. sdNEE / NEE}

\item{nRec}{integer scalar: number of finite observations}

\item{ctrl}{list of further controls, with entry
\code{isLasslopPriorsApplied}}
}
\value{
a numeric vector with prior estimates of the parameters
}
\description{
LRC_getPriorScale
}
\details{
The beta parameter is quite well defined. Hence use a prior with
a standard deviation. The specific results are sometimes a bit sensitive
to the uncertainty of the beta prior.

This uncertainty is set corresponding to 20 times the median relative flux
uncertainty. The prior is weighted n times the observations in the cost.
Hence, overall it is using a weight of 1 / 20 of the weight of all
observations.

However, its not well defined if PAR does not reach saturation. Need to
check before applying this prior
}
