% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_base.R
\name{LRC_isParameterInBounds}
\alias{LRC_isParameterInBounds}
\title{isParameterInBounds}
\usage{
LRC_isParameterInBounds(theta, sdTheta, RRefNight, ctrl)
}
\arguments{
\item{theta}{numeric vector of parameters}

\item{sdTheta}{numeric vector of standard deviation of parameters}

\item{RRefNight}{numeric scalar: night-time based estimate of basal respiration}

\item{ctrl}{list of further controls}
}
\value{
logical scalar: TRUE if parameters are within bounds, FALSE otherwise
}
\description{
Check if estimated parameter vector is within reasonable bounds.
Check the Beta bounds that depend on uncertainty: outside if (beta > 100 and
sdBeta >= beta)
}
