% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_nonrectangular.R
\name{LRC_NonRect_getOptimizedParamPos}
\alias{LRC_NonRect_getOptimizedParamPos}
\title{Get the positions of the parameters to optimize for given Fixed}
\usage{
LRC_NonRect_getOptimizedParamPos(isUsingFixedVPD, isUsingFixedAlpha)
}
\arguments{
\item{isUsingFixedVPD}{boolean: if TRUE, VPD effect set to zero and is not optimized}

\item{isUsingFixedAlpha}{boolean: if TRUE, initial slope is fixed and is not optimized}
}
\value{
integer vector of positions of the parameters to optimize
}
\description{
Get the positions of the parameters to optimize for given Fixed
}
