% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_nonrectangular.R
\name{LRC_NonRect_getPriorScale}
\alias{LRC_NonRect_getPriorScale}
\title{Get the prior distribution of parameters}
\usage{
LRC_NonRect_getPriorScale(thetaPrior, medianRelFluxUncertainty, nRec, ctrl)
}
\arguments{
\item{thetaPrior}{numeric vector of location of priors}

\item{medianRelFluxUncertainty}{numeric scalar: median across the
relative uncertainty of the flux values, i.e. sdNEE / NEE}

\item{nRec}{integer scalar: number of finite observations}

\item{ctrl}{list of further controls, with entry
\code{isLasslopPriorsApplied}}
}
\description{
Get the prior distribution of parameters
}
\keyword{internal}
