% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_base_optim.R
\name{LRC_optimLRC}
\alias{LRC_optimLRC}
\title{optim LRC}
\usage{
LRC_optimLRC(theta, iOpt, sdParameterPrior, ..., ctrl, isUsingHessian)
}
\arguments{
\item{theta}{numeric vector of parameters}

\item{iOpt}{integer vector of positions of parameters that are optimized}

\item{sdParameterPrior}{numeric vector of standard deviation of parameterPrior}

\item{...}{other arguments to \code{\link[=LRC_computeCost]{LRC_computeCost()}}}

\item{ctrl}{list of further controls}

\item{isUsingHessian}{scalar boolean: set to TRUE to compute Hessian at optimum}
}
\value{
list of result of \code{\link[=optim]{optim()}} amended with list
}
\description{
optim LRC
}
