% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_base.R
\name{LRC_optimLRCBounds}
\alias{LRC_optimLRCBounds}
\title{Optimize parameters with refitting with some fixed parameters if outside bounds}
\usage{
LRC_optimLRCBounds(
  theta0,
  parameterPrior,
  ...,
  dsDay,
  lastGoodParameters,
  ctrl
)
}
\arguments{
\item{theta0}{numeric vector of initial parameter estimate}

\item{parameterPrior}{numeric vector of prior estimate of model parameters}

\item{...}{further parameters to \code{.optimLRC}}

\item{dsDay}{dataframe of NEE, sdNEE and predictors Rg, VPD and Temp}

\item{lastGoodParameters}{numeric vector of last successful fit}

\item{ctrl}{list of further controls, such as \code{isNeglectVPDEffect = TRUE}}
}
\value{
list of result of \code{\link[=LRC_optimLRCOnAdjustedPrior]{LRC_optimLRCOnAdjustedPrior()}}
\itemize{
\item \code{theta}: vector of optimized parameters
\item \code{iOpt}: vector of positions of parameters that are optimized
\item \code{thetaInitialGuess}: initial guess from data
see \code{\link[=LRC_fitLRC]{LRC_fitLRC()}}
}
}
\description{
Optimize parameters with refitting with some fixed parameters if outside bounds
}
\details{
If parameters alpha or k are outside bounds (Table A1 in Lasslop 2010),
refit with some parameters fixed to values from fit of previous window.
}
\seealso{
\code{\link[=LRC_fitLRC]{LRC_fitLRC()}}
}
