% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_base.R
\name{LRC_optimLRCOnAdjustedPrior}
\alias{LRC_optimLRCOnAdjustedPrior}
\title{optimLRCOnAdjustedPrior}
\usage{
LRC_optimLRCOnAdjustedPrior(theta, iOpt, dsDay, parameterPrior, ctrl, ...)
}
\arguments{
\item{theta}{numeric vector of starting values}

\item{iOpt}{integer vector: positions of subset of parameters that are optimized}

\item{dsDay}{dataframe of NEE, sdNEE and predictors Rg, VPD and Temp}

\item{parameterPrior}{numeric vector of prior parameter estimates (corresponding to theta)}

\item{ctrl}{list of further controls}

\item{...}{further arguments to
\code{\link[=LRC_optimLRC]{LRC_optimLRC()}} (passed to
\code{\link[=LRC_computeCost]{LRC_computeCost()}})}
}
\value{
list of result of \code{\link[=LRC_optimLRC]{LRC_optimLRC()}} amended with list
\code{theta}, \code{iOpt} and \code{convergence}
}
\description{
Lower bound flux uncertainty and adjust prior uncertainty before calling optimLRC
}
\details{
Only those records are used for optimization where both NEE and sdNEE are finite.
In larger settings, already filtered at \code{partGLFitLRCOneWindow}

Optimization of LRC parameters takes into account the uncertainty of the flux
values. In order to avoid very strong leverage, values with a very low
uncertainty (< a lower quantile) are assigned the lower quantile is assigned.
This procedure downweighs records with a high uncertainty, but does not apply
a large leverage for records with a very low uncertainty. Avoid this
correction by setting \code{ctrl$isBoundLowerNEEUncertainty = FALSE}
}
