% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_base_optim.R
\name{LRC_predictGPP}
\alias{LRC_predictGPP}
\title{Light Response function for GPP}
\usage{
LRC_predictGPP(Rg, ...)
}
\arguments{
\item{Rg}{ppfd -> photosynthetic flux density \link{mumol / m2 / s} or Global Radiation}

\item{...}{further parameters to the LRC}
}
\value{
numeric vector of length(Rg) of GPP
}
\description{
\itemize{
\item Rectangular: \code{\link[=RectangularLRCFitter_predictGPP]{RectangularLRCFitter_predictGPP()}}
\item Nonrectangular: \code{\link[=NonrectangularLRCFitter_predictGPP]{NonrectangularLRCFitter_predictGPP()}}
\item Logistic Sigmoid: \code{\link[=LogisticSigmoidLRCFitter_predictGPP]{LogisticSigmoidLRCFitter_predictGPP()}}
}
}
\seealso{
\code{\link[=partitionNEEGL]{partitionNEEGL()}}
}
