\name{NonrectangularLRCFitter}
\alias{NonrectangularLRCFitter}
\title{NonrectangularLRCFitter}

\description{
Constructs an instance of class \code{\link{NonrectangularLRCFitter-class}}
}

\usage{NonrectangularLRCFitter(...)}
\arguments{
  \item{\dots}{not used.}
}



\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





