% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_nonrectangular.R
\name{NonrectangularLRCFitter_computeGPPGradient}
\alias{NonrectangularLRCFitter_computeGPPGradient}
\title{Gradient of \code{\link[=NonrectangularLRCFitter_predictGPP]{NonrectangularLRCFitter_predictGPP()}}}
\usage{
NonrectangularLRCFitter_computeGPPGradient(Rg, Amax, alpha, logitconv)
}
\arguments{
\item{Rg}{photosynthetic flux density \link{umol / m2 / s} or Global Radiation}

\item{Amax}{saturation (beta parameter) adjusted for effect of VPD}

\item{alpha}{slope at Rg = 0}

\item{logitconv}{-> logit-transformed convexity parameter}
}
\description{
Gradient of \code{\link[=NonrectangularLRCFitter_predictGPP]{NonrectangularLRCFitter_predictGPP()}}
}
