% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_nonrectangular.R
\name{NonrectangularLRCFitter_computeLRCGradient}
\alias{NonrectangularLRCFitter_computeLRCGradient}
\title{Gradient of \code{\link[=NonrectangularLRCFitter_predictLRC]{NonrectangularLRCFitter_predictLRC()}}}
\usage{
NonrectangularLRCFitter_computeLRCGradient(
  theta,
  Rg,
  VPD,
  Temp,
  VPD0 = 10,
  fixVPD = (k == 0),
  TRef = 15
)
}
\arguments{
\item{theta}{-> parameter vector (
theta\link{1} = kVPD (k),
theta\link{2} = beta0 (beta),
theta\link{3} = alpha,
theta\link{4} = RRef (rb), # theta\link{4} = E0,
theta\link{5} = logitconv)}

\item{Rg}{photosynthetic flux density \link{umol / m2 / s} or Global Radiation}

\item{VPD}{Vapor Pressure Deficit \link{hPa}}

\item{Temp}{Temperature \link{degC}}

\item{VPD0}{\link{hPa} -> Parameters VPD0 fixed to 10 hPa according to
Lasslop et al 2010}

\item{fixVPD}{boolean scalar or vector of nrow(theta):
fixVPD if TRUE the VPD effect is not considered and VPD is not part
of the computation}

\item{TRef}{numeric scalar of Temperature (degree Celsius) for reference
respiration \code{RRef}}
}
\description{
Gradient of \code{\link[=NonrectangularLRCFitter_predictLRC]{NonrectangularLRCFitter_predictLRC()}}
}
\details{
This function computes the gradient of the Nonrectangular Light Response
Curve function with respect to the parameters.

Differs from base by extracting \code{conv} parameter from \code{theta} and adding
gradient to \code{logitconv} (3rd parameter from computeGPPGradient).
}
