% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_nonrectangular.R
\name{NonrectangularLRCFitter_getOptimizedParameterPositions}
\alias{NonrectangularLRCFitter_getOptimizedParameterPositions}
\title{Get the positions of the parameters to optimize for given Fixed}
\usage{
NonrectangularLRCFitter_getOptimizedParameterPositions(
  isUsingFixedVPD,
  isUsingFixedAlpha
)
}
\arguments{
\item{isUsingFixedVPD}{boolean scalar: if TRUE, VPD effect set to zero and is not optimized}

\item{isUsingFixedAlpha}{boolean scalar: if TRUE, initial slope is fixed and is not optimized}
}
\value{
integer vector of positions of the parameters to optimize
}
\description{
Get the positions of the parameters to optimize for given Fixed
}
