% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitioningLasslop10Nighttime.R
\name{partGL_FitNight_1win_E0}
\alias{partGL_FitNight_1win_E0}
\title{Estimate temperature sensitivity E0 within one window}
\usage{
partGL_FitNight_1win_E0(
  dss,
  winInfo,
  prevRes,
  isVerbose = TRUE,
  nRecInDay = 48L,
  controlGLPart = partGLControl()
)
}
\arguments{
\item{dss}{data.frame with numeric columns NEE, sdNEE, Temp (degC) , VPD, Rg,
and logical columns isNight and isDay}

\item{winInfo}{one-row data.frame with window information, including iWindow}

\item{prevRes}{component prevRes from previous result, here with item prevE0}

\item{isVerbose}{set to FALSE to suppress messages}

\item{nRecInDay}{number of records within one day (for half-hourly data its 48)}

\item{controlGLPart}{list of further default parameters}
}
\description{
Estimate temperature sensitivity E0 within one window
}
\details{
Estimation of respiration at reference temperature (RRef) and temperature
sensitivity (E0) for one window of night-time data.

The reference temperature is taken as the median of the temperatures of
valid records, unless a fixed reference temperature (in degree Celsius) is
specified by argument \code{controlGLPart$fixedTRefAtNightTime}.
}
\seealso{
\code{\link[=partitionNEEGL]{partitionNEEGL()}}, \code{\link[=partGL_FitNight_1win_E0_kernel]{partGL_FitNight_1win_E0_kernel()}}
}
