% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitioningLasslop10Nighttime.R
\name{partGLFitNightTimeTRespSens}
\alias{partGLFitNightTimeTRespSens}
\title{Partitioning of NEE into GPP and RE using the Lasslop et al. (2010) method}
\usage{
partGLFitNightTimeTRespSens(
  ds,
  winSizeRefInDays = 4L,
  winSizeNight = 12L,
  winExtendSizes = winSizeNight * c(2L, 4L),
  strideInDays = 2L,
  isVerbose = TRUE,
  nRecInDay = 48L,
  controlGLPart = partGLControl()
)
}
\arguments{
\item{ds}{data.frame with numeric columns NEE, sdNEE, Temp (degC),
VPD, Rg, and logical columns isNight and isDay}

\item{winSizeRefInDays}{Window size in days for daytime for referencing
windows to the same base}

\item{winSizeNight}{Window size in days for nighttime fits}

\item{winExtendSizes}{successively increased nighttime windows, to obtain a
night-time fit}

\item{strideInDays}{step in days for shifting the windows}

\item{isVerbose}{set to FALSE to suppress messages}

\item{nRecInDay}{number of records within one day (for half-hourly data its 48)}

\item{controlGLPart}{list of further default parameters}
}
\description{
Estimate temperature sensitivity parameters for successive periods
}
\details{
Window sizes are successively increased to \code{winExtendSizes} in order to obtain
parameter estimates where fits with smaller window size failed. This behaviour
of repeated attempts can be avoid by setting \code{controlGLPart$isExtendTRefWindow = FALSE}.
}
\seealso{
\code{\link[=partGLFitNightTempSensOneWindow]{partGLFitNightTempSensOneWindow()}}, \code{\link[=partGLSmoothTempSens]{partGLSmoothTempSens()}},
\code{\link[=partGLFitNightRespRefOneWindow]{partGLFitNightRespRefOneWindow()}}
}
