% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitioningLasslop10Nighttime.R
\name{partGLSmoothTempSens}
\alias{partGLSmoothTempSens}
\title{Smoothes time development of E0}
\usage{
partGLSmoothTempSens(E0Win)
}
\arguments{
\item{E0Win}{data.frame with columns E0 and sdE0, RRefFit, and TRefFit
with one row for each window}
}
\description{
Smoothes time development of E0
}
