% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileHandlingFormats.R
\name{posix_to_isostring}
\alias{posix_to_isostring}
\alias{isostring_to_posix}
\title{convert between timestamp formats}
\usage{
posix_to_isostring(x)

isostring_to_posix(x)
}
\arguments{
\item{x}{vector of timestamps in input format}
}
\value{
fector of timestamps in output format
}
\description{
convert between timestamp formats
}
\section{Functions}{
\itemize{
\item \code{posix_to_isostring}: convert POSIX format to ISO string format used in Fluxnet15: YYYYMMDDHHMM

\item \code{isostring_to_posix}: convert ISO string format used in Fluxnet15: YYYYMMDDHHMM to POSIXct
}}

