% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileHandlingFormats.R
\name{read_from_ameriflux22}
\alias{read_from_ameriflux22}
\title{Extract basic variables from Ameriflux standard (as of 2022) data.frames}
\usage{
read_from_ameriflux22(df)
}
\arguments{
\item{df}{data.frame: with columns FC, SW_IN, RH, TA, USTAR, L and E}
}
\value{
Data.Frame with columns 
  DateTime, NEE,	Rg,	Tair,	rH,	VPD, Ustar, LE, H
}
\description{
NEE is read from FC, Rg from SW_in, VPD is computed from RH and Tair.
Non-storage corrected LE and H are read.
}
