% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_rectangular.R
\docType{class}
\name{RectangularLRCFitter-class}
\alias{RectangularLRCFitter-class}
\alias{RectangularLRCFitter}
\title{R5 reference class for the Rectangular Light response curve}
\description{
R5 reference class for the Rectangular Light response curve
}

