% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_rectangular.R
\name{RectangularLRCFitter_computeGPPGradient}
\alias{RectangularLRCFitter_computeGPPGradient}
\title{RectangularLRCFitter_computeGPPGradient}
\usage{
RectangularLRCFitter_computeGPPGradient(Rg, Amax, alpha)
}
\arguments{
\item{Rg}{[numeric] ppfd -> photosynthetic flux density [mumol / m2 / s] or Global Radiation}

\item{Amax}{[numeric] saturation (beta parameter) adjusted for effect of VPD}

\item{alpha}{[numeric] slope at Rg = 0}
}
\value{
numeric matrix (length(Rg), 2) of gradients of predicted GPP
to Amax and alpha
}
\description{
RectangularLRCFitter_computeGPPGradient
}
