% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_rectangular.R
\docType{class}
\name{RectangularLRCFitterCVersion-class}
\alias{RectangularLRCFitterCVersion-class}
\alias{RectangularLRCFitterCVersion}
\title{R5 reference class C-version for the Rectangular Light response curve}
\description{
R5 reference class C-version for the Rectangular Light response curve
}

