\name{RectangularLRCFitterCVersion}
\alias{RectangularLRCFitterCVersion}
\title{RectangularLRCFitterCVersion}

\description{
Constructs an instance of class \code{\link{RectangularLRCFitterCVersion-class}}
}
\usage{RectangularLRCFitterCVersion(...)}
\arguments{
  \item{\dots}{not used.}
}



\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





