% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_rectangular.R
\name{RectangularLRCFitterCVersion_computeCost}
\alias{RectangularLRCFitterCVersion_computeCost}
\title{Computing residual sum of squares for predictions vs. data of NEE}
\usage{
RectangularLRCFitterCVersion_computeCost(
  thetaOpt,
  theta,
  iOpt,
  flux,
  sdFlux,
  parameterPrior,
  sdParameterPrior,
  ...,
  VPD0 = 10,
  fixVPD = (k == 0)
)
}
\arguments{
\item{thetaOpt}{parameter vector with components of theta0 that are optimized}

\item{theta}{parameter vector with positions as in argument of
\code{\link[=LightResponseCurveFitter_getParameterNames]{LightResponseCurveFitter_getParameterNames()}}}

\item{iOpt}{position in theta that are optimized}

\item{flux}{numeric: NEP (-NEE) or GPP time series \link{umolCO2 / m2 / s},
should not contain NA}

\item{sdFlux}{numeric: standard deviation of Flux \link{umolCO2 / m2 / s},
should not contain NA}

\item{parameterPrior}{numeric vector along theta: prior estimate of
parameter (range of values)}

\item{sdParameterPrior}{standard deviation of parameterPrior}

\item{...}{other arguments to
\link{LightResponseCurveFitter_predictLRC}, such as VPD0, fixVPD}

\item{VPD0}{VPD0 \link{hPa} -> Parameters VPD0 fixed to 10 hPa according
to Lasslop et al 2010}

\item{fixVPD}{boolean scalar or vector of nrow theta:fixVPD
if TRUE the VPD effect is not considered and VPD is not
part of the computation}
}
\description{
Computing residual sum of squares for predictions vs. data of NEE
}
