\name{REddyProc-package}
\alias{REddyProc-package}
\docType{package}
\title{Package for data processing of half-hourly eddy-covariance measurements}
\description{R eddy processing package with MDS gap filling algorithm adopted after PV-Wave source code from Markus Reichstein. Package is being extended to also include ustar filtering and flux partitioning.}
\details{
\tabular{ll}{Package: \tab REddyProc\cr
Title: \tab Package for data processing of half-hourly eddy-covariance measurements\cr
Version: \tab 0.31\cr
Date: \tab 2013-03-31\cr
Author: \tab (Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)\cr
Maintainer: \tab Antje M. Moffat <antje.moffat@bgc.mpg.de>\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 1.15.0), methods\cr
Suggests: \tab RNetCDF, testthat, inlinedocs\cr}
}
\author{(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}

\keyword{ package }


