% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileHandling.R
\name{REddyProc_defaultunits}
\alias{REddyProc_defaultunits}
\title{Get the default units for given variables}
\usage{
REddyProc_defaultunits(variable_names)
}
\arguments{
\item{variable_names}{string vector of variables to query units for}
}
\value{
string vector with units, NA for non-standard variables.
}
\description{
Get the default units for given variables
}
