\name{renameVariablesInDataframe}
\alias{renameVariablesInDataframe}
\title{renameVariablesInDataframe}
\description{Rename the column names of a data.frame according to a given mapping}
\usage{renameVariablesInDataframe(data.F, mapping = getBGC05ToAmerifluxVariableNameMapping())}
\arguments{
  \item{data.F}{data.frame whose columns should be renamped}
  \item{mapping}{named character vector:
<< specifying a renaming (name -> value)  
<< of the variables, see e.g. \code{\link{getAmerifluxToBGC05VariableNameMapping}}}
}



\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





