\name{sEddyProc_sExportResults}
\alias{sEddyProc_sExportResults}
\title{sEddyProc$sExportData - Export internal sTEMP data frame with result columns}
\description{Export class internal sTEMP data frame with result columns}
\usage{sEddyProc_sExportResults(isListColumnsExported = FALSE)}
\arguments{
  \item{isListColumnsExported}{if TRUE export list columns in addition to numeric columns,
such as the covariance matrices of the the day-time-partitioning LRC fits}
}

\value{Return data frame sTEMP with results.}

\author{AMM
Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut Šigut <sigut.l@czechglobe.cz> [ctb]}





