% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_vpd_from_dew.R
\name{sEddyProc_sFillVPDFromDew}
\alias{sEddyProc_sFillVPDFromDew}
\title{Estimate VPD from daily minimum temperature}
\usage{
sEddyProc_sFillVPDFromDew(...)
}
\arguments{
\item{...}{further arguments to \code{\link{estimate_vpd_from_dew}}}
}
\value{
side effect of updated column VPDfromDew in class
}
\description{
of the data in the class function using \code{\link{estimate_vpd_from_dew}}.
}
