\name{sEddyProc_sGetData}
\alias{sEddyProc_sGetData}
\title{sEddyProc$sGetData - Get internal sDATA data frame}
\description{Get class internal sDATA data frame}
\usage{sEddyProc_sGetData()}

\value{Return data frame sDATA.}

\author{AMM
Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





