\name{sEddyProc_sGetUstarScenarios}
\alias{sEddyProc_sGetUstarScenarios}
\title{sEddyProc sGetUstarScenarios}
\description{get the current uStar processing scenarios}
\usage{sEddyProc_sGetUstarScenarios()}
\details{the associated suffixes can be retrieved by
\code{colnames(myClass$sGetUstarScenarios())[-1]}}
\value{a data.frame with first column listing each season and
other column a scenario of uStar thresholds.}

\author{Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]}



\seealso{\code{\link{sEddyProc_sSetUstarScenarios}}}

