\name{sEddyProc_sPlotHHFluxesY}
\alias{sEddyProc_sPlotHHFluxesY}
\title{sEddyProc$sPlotHHFluxesY -  Plot half-hourly fluxes of specified year}
\description{The half-hourly fluxes for a single year are plotted to the current device,
scaled to all data.}
\usage{sEddyProc_sPlotHHFluxesY(Var.s, QFVar.s = "none", 
    QFValue.n = NA_real_, Year.i)}
\arguments{
  \item{Var.s}{Variable to plot}
  \item{QFVar.s}{Quality flag of variable to be filled}
  \item{QFValue.n}{Value of quality flag for data to plot}
  \item{Year.i}{Year to plot}
}



\author{AMM, KS
Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut Šigut <sigut.l@czechglobe.cz> [ctb]}





