\name{sEddyProc_sPrintFrames}
\alias{sEddyProc_sPrintFrames}
\title{sEddyProc$sPrintFrames - Print internal sDATA and sTEMP data frame}
\description{Print class internal sDATA and sTEMP data frame}
\usage{sEddyProc_sPrintFrames(NumRows.i = 100)}
\arguments{
  \item{NumRows.i}{Number of rows to print}
}

\value{Print the first rows of class internal sDATA and sTEMP data frame.}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





