\name{sEddyProc_sSetUStarSeasons}
\alias{sEddyProc_sSetUStarSeasons}
\title{sEddyProc sSetUStarSeasons}
\description{Defining seasons for the uStar threshold estimation}
\usage{sEddyProc_sSetUStarSeasons(seasonFactor = usCreateSeasonFactorMonth(sDATA$sDateTime))}
\arguments{
  \item{seasonFactor}{factor for subsetting times with different uStar threshold (see details)}
}

\value{class with updated \code{seasonFactor}}

\author{TW
Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





