\name{sEddyProc_sTKFluxPartition}
\alias{sEddyProc_sTKFluxPartition}
\title{sEddyProc sTKFluxPartition}
\description{Modified daytime-based Flux partitioning after Keenan et al. (2019)}
\usage{sEddyProc_sTKFluxPartition(..., controlGLPart = partGLControl())}
\arguments{
  \item{\dots}{arguments to \code{\link{sEddyProc_sGLFluxPartition}}
in addition to the dataset}
  \item{controlGLPart}{further default parameters,
such as \code{suffix}}
}

\value{Flux partitioning results are in sTEMP data frame of the class.}

\author{Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





