\name{sEddyProc_useAnnualUStarThresholds}
\alias{sEddyProc_useAnnualUStarThresholds}
\title{sEddyProc useAnnualUStarThresholds}
\description{use seasonal estimates of uStar thresholds}
\usage{sEddyProc_useAnnualUStarThresholds()}



\author{Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]}



\seealso{\code{\link{sEddyProc_sSetUstarScenarios}},
\code{\link{sEddyProc_useSeaonsalUStarThresholds}}}

