\name{sExportResults}
\alias{sExportResults}
\title{sEddyProc$sExportData - Export internal sTEMP data frame with result columns}
\description{Export class internal sDATA data frame}
\usage{sExportResults()}

\value{Return data frame sTEMP with results.}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





