\name{sFillMDC}
\alias{sFillMDC}
\title{sEddyProc$sFillMDC - Gap filling with Mean Diurnal Course (MDC)}
\description{Mean Diurnal Course (MDC) algorithm based on average values within +/- one hour of adjacent days}
\usage{sFillMDC(WinDays.i, Verbose.b = TRUE)}
\arguments{
  \item{WinDays.i}{Window size for filling in days}
  \item{Verbose.b}{Print status information to screen}
}

\value{MDC filling results in sTEMP data frame.}

\author{AMM 
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





