\name{sGetData}
\alias{sGetData}
\title{sEddyProc$sGetData - Get internal sDATA data frame}
\description{Get class internal sDATA data frame}
\usage{sGetData()}

\value{Return data frame sDATA.}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





