% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitioningLasslop10Nighttime.R
\name{simplifyApplyWindows}
\alias{simplifyApplyWindows}
\title{simplify the result returned by applyWindows}
\usage{
simplifyApplyWindows(resApply)
}
\arguments{
\item{resApply}{result of \code{\link[=applyWindows]{applyWindows()}}}
}
\value{
A single data.frame with columns of winInfo and results of FUN
}
\description{
simplify the result returned by applyWindows
}
\details{
If FUN returns a named vector or a single-row data.frame,
the resFUN result component of applyWindows can be condensed to a data.frame.
This result is column-bound to the winInfo result component
}
