\name{sPlotDailySumsY}
\alias{sPlotDailySumsY}
\title{sEddyProc$sPlotDailySumsY - Plot daily sum of specified year}
\description{The daily sums for a single year are plotted to the current device, scaled to all data.
The daily sums are only calculated for days with complete data.}
\usage{sPlotDailySumsY(Var.s, VarUnc.s, Year.i)}
\arguments{
  \item{Var.s}{(Filled) variable to plot}
  \item{VarUnc.s}{Uncertainty estimates for variable}
  \item{Year.i}{Year to plot}
}



\author{KS, AMM
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





