\name{sPlotDiurnalCycle}
\alias{sPlotDiurnalCycle}
\title{sEddyProc$sPlotDiurnalCycle - Image with diurnal cycles of each month}
\description{Generates image in specified format ('pdf' or 'png') with diurnal cycles.}
\usage{sPlotDiurnalCycle(Var.s, QFVar.s = "none", QFValue.n = NA_real_, 
    Format.s = "pdf", Dir.s = "plots")}
\arguments{
  \item{Var.s}{Variable to plot}
  \item{QFVar.s}{Quality flag of variable to be filled}
  \item{QFValue.n}{Value of quality flag for data to plot}
  \item{Format.s}{Graphics file format ('pdf' or 'png')}
  \item{Dir.s}{Directory for plotting}
}



\author{KS, AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





