\name{sPlotDiurnalCycleM}
\alias{sPlotDiurnalCycleM}
\title{sEddyProc$sPlotDiurnalCycleM - Plot diurnal cycles of specified month}
\description{The diurnal cycles of a single month are potted to the current device, scaled to all data. Each year is plotted as a different (coloured) line.}
\usage{sPlotDiurnalCycleM(Var.s, QFVar.s = "none", QFValue.n = NA_real_, 
    Month.i, Legend.b = T)}\arguments{
  \item{Var.s}{Variable to plot}
  \item{QFVar.s}{Quality flag of variable to be filled}
  \item{QFValue.n}{Value of quality flag for data to plot}
  \item{Month.i}{Month to plot}
  \item{Legend.b}{Plot with legend}
}



\author{AMM, KS
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





