\name{sPlotFingerprintY}
\alias{sPlotFingerprintY}
\title{sEddyProc$sPlotFingerprintY - Plot fingerprint of specified year}
\description{The fingerprint for a single year is plotted to the current device, scaled to all data.}
\usage{sPlotFingerprintY(Var.s, QFVar.s = "none", QFValue.n = NA_real_, 
    Year.i, Legend.b = F)}
\arguments{
  \item{Var.s}{Variable to plot}
  \item{QFVar.s}{Quality flag of variable to be filled}
  \item{QFValue.n}{Value of quality flag for data to plot}
  \item{Year.i}{Year to plot}
  \item{Legend.b}{Plot only legend}
}



\author{AMM, KS
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





