\name{sPlotHHFluxesY}
\alias{sPlotHHFluxesY}
\title{sEddyProc$sPlotHHFluxesY -  Plot half-hourly fluxes of specified year}
\description{The half-hourly fluxes for a single year are plotted to the current device, scaled to all data.}
\usage{sPlotHHFluxesY(Var.s, QFvar.s, QFvalue.n, Year.i)}
\arguments{
  \item{Var.s}{Variable to plot}
  \item{QFvar.s}{Quality flag of variable to be filled}
  \item{QFvalue.n}{Value of quality flag for data to plot}
  \item{Year.i}{Year to plot}
}



\author{KS, AMM
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





