\name{sPlotHHMeans}
\alias{sPlotHHMeans}
\title{sEddyProc$sPlotHHMeans - Image with half-hourly means of each month}
\description{Generates image in specified format ('pdf' or 'png') with half-hourly means, see also \code{\link{sPlotHHMeansM}}.}
\usage{sPlotHHMeans(Var.s, QFvar.s = "none", QFvalue.n = NA, Format.s = "pdf", 
    Dir.s = "plots")}
\arguments{
  \item{Var.s}{Variable to plot}
  \item{QFvar.s}{Quality flag of variable to be filled}
  \item{QFvalue.n}{Value of quality flag for data to plot}
  \item{Format.s}{Graphics file format ('pdf' or 'png')}
  \item{Dir.s}{Directory for plotting}
}



\author{KS, AMM
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





