\name{sPlotHHMeansM}
\alias{sPlotHHMeansM}
\title{sEddyProc$sPlotHHMeansM - Plot half-hourly means of specified months}
\description{The half-hourly means of a single months (averaged over all years) are potted to the current device, scaled to all data.}
\usage{
sPlotHHMeansM(Var.s,
QFvar.s,
QFvalue.n,
Month.i)
}
\arguments{
  \item{Var.s}{Variable to plot}
  \item{QFvar.s}{Quality flag of variable to be filled}
  \item{QFvalue.n}{Value of quality flag for data to plot}
  \item{Month.i}{Month to plot}
}



\author{KS, AMM
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





