\name{sxClosePlot}
\alias{sxClosePlot}
\title{sEddyProc - Internal function}
\description{Close screens and save graphics device to file.}
\usage{sxClosePlot(PlotFile.s)}
\arguments{
  \item{PlotFile.s}{Name of opened graphics device}
}



\author{KS, AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





