\name{sxSetTitle}
\alias{sxSetTitle}
\title{sEddyProc - Internal function}
\description{Set title of plot.}
\usage{sxSetTitle(Var.s, QFvar.s, QFvalue.n, Name.s)}
\arguments{
  \item{Var.s}{Variable to plot}
  \item{QFvar.s}{Quality flag of variable to be filled}
  \item{QFvalue.n}{Value of quality flag for data to plot}
  \item{Name.s}{Name of plot}
}

\value{String with plot title}

\author{KS, AMM
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





