\name{usGetAnnualSeasonUStarMap}
\alias{usGetAnnualSeasonUStarMap}
\title{usGetAnnualSeasonUStarMap}
\description{extract mapping season -> uStar columns from Distribution result }
\usage{usGetAnnualSeasonUStarMap(uStarTh)}
\arguments{
  \item{uStarTh}{result of \code{\link{sEddyProc_sEstUstarThresholdDistribution}} or \code{\link{sEddyProc_sEstUstarThreshold}}$uStarTh}
}



\author{TW
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





