\name{usGetAnnualSeasonUStarMappingFromDistributionResult}
\alias{usGetAnnualSeasonUStarMappingFromDistributionResult}
\title{usGetAnnualSeasonUStarMappingFromDistributionResult}
\description{extract mapping season -> uStar columns from Distribution result }
\usage{usGetAnnualSeasonUStarMappingFromDistributionResult(uStarTh)}
\arguments{
  \item{uStarTh}{result of \code{\link{sEstUstarThresholdDistribution}} or \code{\link{sEstUstarThreshold}}$uStarTh}
}



\author{TW
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





