\name{usGetSeasonalSeasonUStarMappingFromDistributionResult}
\alias{usGetSeasonalSeasonUStarMappingFromDistributionResult}
\title{usGetSeasonalSeasonUStarMappingFromDistributionResult}
\description{extract mapping season -> uStar columns from Distribution result (\code{\link{sEstUstarThresholdDistribution}}}
\usage{usGetSeasonalSeasonUStarMappingFromDistributionResult(uStarTh)}
\arguments{
  \item{uStarTh}{result of \code{\link{sEstUstarThresholdDistribution}} or \code{\link{sEstUstarThreshold}}$uStarTh}
}

\value{a data frame with first column the season, and other columns different uStar threshold estimates}

\author{TW
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





