\name{whichValueGreaterEqualC}
\alias{whichValueGreaterEqualC}
\title{whichValueGreaterEqualC}

\usage{whichValueGreaterEqualC(x, threshold, iStart)}\arguments{
  \item{x}{
}
  \item{threshold}{
}
  \item{iStart}{
}
}



\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





