% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitioningLasslop10Nighttime.R
\name{applyWindows}
\alias{applyWindows}
\title{apply a function to several windows of a data.frame}
\usage{
applyWindows(
  ds,
  FUN,
  prevRes = list(),
  winSizeInDays = winSizeRefInDays,
  winSizeRefInDays = 4L,
  strideInDays = floor(winSizeRefInDays/2L),
  isVerbose = TRUE,
  nRecInDay = 48L,
  ...
)
}
\arguments{
\item{ds}{data.frame to iterate}

\item{FUN}{function to apply to subsets of the data.frame taking a subset of
the data.frame as first argument the second: a one-row data.frame with window
information (iWindow, dayStart, dayEnd, iRecStart, iRecEnd, iCentralRec) the
third: most recent valid result of calls FUN. Valid is a non-NULL result.}

\item{prevRes}{initial values for the list that
is carried between windows}

\item{winSizeInDays}{Window size in days}

\item{winSizeRefInDays}{Window size in days for reference window
(e.g. day-Window for night time)}

\item{strideInDays}{step in days for shifting the window, for alligning
usually a factor of winSizeRef}

\item{isVerbose}{set to FALSE to suppress messages}

\item{nRecInDay}{number of records within one day (for half-hourly data its 48)}

\item{...}{further arguments to FUN}
}
\description{
apply a function to several windows of a data.frame
}
