% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileHandlingFormats.R
\name{extract_FN15}
\alias{extract_FN15}
\title{extract processing results with columns corresponding to Fluxnet15 release}
\usage{
extract_FN15(
  EProc = .self,
  is_export_nonfilled = TRUE,
  keep_other_cols = FALSE
)
}
\arguments{
\item{EProc}{sEddyProc class with uncertainty also in meteo variables and
both nighttime and daytime partitioning columns present}

\item{is_export_nonfilled}{set to FALSE to not export columns before gapfilling}

\item{keep_other_cols}{set to TRUE to report also other columns}
}
\value{
data.frame with columns names of Fluxnet15. Timestamps are
  in ISO string format \code{\link{POSIXctToBerkeleyJulianDate}}
}
\description{
extract processing results with columns corresponding to Fluxnet15 release
}
