\name{fCalcAVPfromVMFandPress}
\alias{fCalcAVPfromVMFandPress}
\title{fCalcAVPfromVMFandPress}
\description{Calculate AVP from VMF and Press}
\usage{fCalcAVPfromVMFandPress(VMF.V.n, Press.V.n)}
\arguments{
  \item{VMF.V.n}{Vapor mole fraction (VMF, mol / mol)}
  \item{Press.V.n}{Atmospheric pressure (Press, hPa)}
}

\value{Data vector of actual vapor pressure (AVP, hPa (mbar))}

\author{AMM
Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





