\name{fCalcAVPfromVMFandPress}
\alias{fCalcAVPfromVMFandPress}
\title{fCalcAVPfromVMFandPress}
\description{Calculate AVP from VMF and Press}
\usage{fCalcAVPfromVMFandPress(VMF = VMF.V.n, Press = Press.V.n, 
    VMF.V.n, Press.V.n)}
\arguments{
  \item{VMF}{Vapor mole fraction (VMF, mol / mol)}
  \item{Press}{Atmospheric pressure (Press, hPa)}
  \item{VMF.V.n}{deprecated}
  \item{Press.V.n}{deprecated}
}

\value{Data vector of actual vapor pressure (AVP, hPa (mbar))}

\author{AMM
Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





