\name{fCalcETfromLE}
\alias{fCalcETfromLE}
\title{fCalcETfromLE}
\description{Calculate ET from LE and Tair}
\usage{fCalcETfromLE(LE.V.n, Tair.V.n)}
\arguments{
  \item{LE.V.n}{Data vector of latent heat (LE, W m-2)}
  \item{Tair.V.n}{Data vector of air temperature (Tair, degC)}
}

\value{Data vector of evapotranspiration (ET, mmol H20 m-2 s-1)}

\author{AMM
Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





