\name{fCalcETfromLE}
\alias{fCalcETfromLE}
\title{fCalcETfromLE}
\description{Calculate ET from LE and Tair}
\usage{fCalcETfromLE(LE = LE.V.n, Tair = Tair.V.n, 
    LE.V.n, Tair.V.n)}
\arguments{
  \item{LE}{Data vector of latent heat (LE, W m-2)}
  \item{Tair}{Data vector of air temperature (Tair, degC)}
  \item{LE.V.n}{deprecated}
  \item{Tair.V.n}{deprecated}
}

\value{Data vector of evapotranspiration (ET, mmol H20 m-2 s-1)}

\author{AMM
Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]}





