\name{fCalcExtRadiation}
\alias{fCalcExtRadiation}
\title{fCalcExtRadiation}
\description{Calculate the extraterrestrial solar radiation with the eccentricity correction }
\usage{fCalcExtRadiation(DoY.V.n)}
\arguments{
  \item{DoY.V.n}{Data vector with day of year (DoY)}
}

\value{Data vector of extraterrestrial radiation (ExtRad, W_m-2)}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





