\name{fCalcExtRadiation}
\alias{fCalcExtRadiation}
\title{fCalcExtRadiation}
\description{Calculate the extraterrestrial solar radiation with the
eccentricity correction}
\usage{fCalcExtRadiation(DoY = DoY.V.n, DoY.V.n)}
\arguments{
  \item{DoY}{
}
  \item{DoY.V.n}{Data vector with day of year (DoY)}
}

\value{Data vector of extraterrestrial radiation (ExtRad, W_m-2)}

\author{AMM
Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





