\name{fCalcLengthOfGaps}
\alias{fCalcLengthOfGaps}
\title{fCalcLengthOfGaps}
\description{Calculate length of gaps (with flag NA) in specified vector}
\usage{fCalcLengthOfGaps(Data.V.n)}
\arguments{
  \item{Data.V.n}{Numeric vector with gaps (missing values, NAs)}
}

\value{An integer vector with length of gap since start of gap.}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





