\name{fCalcRHfromAVPandTair}
\alias{fCalcRHfromAVPandTair}
\title{Calculate relative humidity from actual vapour pressure and air tempature}

\usage{fCalcRHfromAVPandTair(AVP.V.n, Tair.V.n)}
\arguments{
  \item{AVP.V.n}{Data vector of actual vapour pressure (AVP, mbar)}
  \item{Tair.V.n}{Data vector of air temperature (Tair, °C)}
}

\value{Data vector of relative humidity (rH, \%)}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





