\name{fCalcRHfromAVPandTair}
\alias{fCalcRHfromAVPandTair}
\title{fCalcRHfromAVPandTair}
\description{Calculate relative humidity from actual vapour pressure and air temperature}
\usage{fCalcRHfromAVPandTair(AVP = AVP.V.n, Tair = Tair.V.n, 
    AVP.V.n, Tair.V.n)}
\arguments{
  \item{AVP}{Data vector of actual vapour pressure (AVP, hPa (mbar))}
  \item{Tair}{Data vector of air temperature (Tair, degC)}
  \item{AVP.V.n}{Data vector of actual vapour pressure (AVP, hPa (mbar))}
  \item{Tair.V.n}{Data vector of air temperature (Tair, degC)}
}

\value{Data vector of relative humidity (rH, \%)}

\author{AMM
Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





