\name{fCalcSVPfromTair}
\alias{fCalcSVPfromTair}
\title{Calculate SVP (of water) from Tair}

\usage{fCalcSVPfromTair(Tair.V.n)}
\arguments{
  \item{Tair.V.n}{Data vector of air temperature (Tair, °C)}
}

\value{Data vector of saturation VP (SVP, mbar)}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





