\name{fCalcSVPfromTair}
\alias{fCalcSVPfromTair}
\title{fCalcSVPfromTair}
\description{Calculate SVP (of water) from Tair}
\usage{fCalcSVPfromTair(Tair.V.n)}
\arguments{
  \item{Tair.V.n}{Data vector of air temperature (Tair, degC)}
}

\value{Data vector of saturation vapor pressure (SVP, hPa (mbar))}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





