\name{fCalcSVPfromTair}
\alias{fCalcSVPfromTair}
\title{fCalcSVPfromTair}
\description{Calculate SVP (of water) from Tair}
\usage{fCalcSVPfromTair(Tair = Tair.V.n, Tair.V.n)}
\arguments{
  \item{Tair}{Data vector of air temperature (Tair, degC)}
  \item{Tair.V.n}{deprecated}
}

\value{Data vector of saturation vapor pressure (SVP, hPa (mbar))}

\author{AMM
Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]}





