\name{fCalcSVPfromTair}
\alias{fCalcSVPfromTair}
\title{fCalcSVPfromTair}
\description{Calculate SVP (of water) from Tair}
\usage{fCalcSVPfromTair(Tair.V.n)}
\arguments{
  \item{Tair.V.n}{Data vector of air temperature (Tair, degC)}
}

\value{Data vector of saturation vapor pressure (SVP, hPa (mbar))}

\author{AMM
Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





