\name{fCalcVPDfromRHandTair}
\alias{fCalcVPDfromRHandTair}
\title{Calculate VPD from rH and Tair}

\usage{fCalcVPDfromRHandTair(RH.V.n, Tair.V.n)}
\arguments{
  \item{RH.V.n}{Data vector of relative humidity (rH, \%)}
  \item{Tair.V.n}{Data vector of air temperature (Tair,°C)}
}

\value{Data vector of vapour pressure deficit (VPD, hPa)}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





