\name{fCalcVPDfromRHandTair}
\alias{fCalcVPDfromRHandTair}
\title{fCalcVPDfromRHandTair}
\description{Calculate VPD from rH and Tair}
\usage{fCalcVPDfromRHandTair(RH.V.n, Tair.V.n)}
\arguments{
  \item{RH.V.n}{Data vector of relative humidity (rH, \%)}
  \item{Tair.V.n}{Data vector of air temperature (Tair, degC)}
}

\value{Data vector of vapour pressure deficit (VPD, hPa (mbar))}

\author{AMM
Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





