\name{fCheckColNames}
\alias{fCheckColNames}
\title{fCheckColNames}
\description{Check if specified columns exists in data frame}
\usage{fCheckColNames(Data.F, ColNames.V.s, CallFunction.s = "")}
\arguments{
  \item{Data.F}{Data frame}
  \item{ColNames.V.s}{Vector of variables to be checked}
  \item{CallFunction.s}{Name of function called from}
}

\value{Function stops on errors.}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





