\name{fCheckValNum}
\alias{fCheckValNum}
\title{Check if variable is a numeric}

\usage{fCheckValNum(Value.n)}
\arguments{
  \item{Value.n}{Value to be checked if numeric}
}
\details{See test_CheckValue.R for more details.}
\value{Boolean value if true of false.}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





