\name{fConvertCtoK}
\alias{fConvertCtoK}
\title{Convert degree Celsius to degree Kelvin}

\usage{fConvertCtoK(Celsius.V.n)}
\arguments{
  \item{Celsius.V.n}{Data vector in Celsius (°C)}
}

\value{Data vector in temperature Kelvin (Temp_K, °K)}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





