\name{fConvertCtoK}
\alias{fConvertCtoK}
\title{fConvertCtoK}
\description{Convert degree Celsius to degree Kelvin}
\usage{fConvertCtoK(Celsius.V.n)}
\arguments{
  \item{Celsius.V.n}{Data vector in Celsius (degC)}
}

\value{Data vector in temperature Kelvin (Temp_K, degK)}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





