\name{fConvertCtoK}
\alias{fConvertCtoK}
\title{fConvertCtoK}
\description{Convert degree Celsius to degree Kelvin}
\usage{fConvertCtoK(Celsius = Celsius.V.n, Celsius.V.n)}
\arguments{
  \item{Celsius}{Data vector in Celsius (degC)}
  \item{Celsius.V.n}{deprecated way of specifying Celsius}
}

\value{Data vector in temperature Kelvin (Temp_K, degK)}

\author{AMM
Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]}





