\name{fConvertGlobalToVisible}
\alias{fConvertGlobalToVisible}
\title{fConvertGlobalToVisible}
\description{Partition global (solar) radiation into only visible (the rest is UV and infrared)}
\usage{fConvertGlobalToVisible(Global.V.n)}
\arguments{
  \item{Global.V.n}{Data vector of global radiation (W m-2)}
}

\value{Data vector of visible part of solar radiation (VisRad, W m-2)}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





