\name{fConvertGlobalToVisible}
\alias{fConvertGlobalToVisible}
\title{fConvertGlobalToVisible}
\description{Partition global (solar) radiation into only visible (the rest is UV and infrared)}
\usage{fConvertGlobalToVisible(Global = Global.V.n, 
    Global.V.n)}
\arguments{
  \item{Global}{Data vector of global radiation (W m-2)}
  \item{Global.V.n}{deprecated}
}

\value{Data vector of visible part of solar radiation (VisRad, W m-2)}

\author{AMM
Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





