\name{fConvertKtoC}
\alias{fConvertKtoC}
\title{fConvertKtoC}
\description{Convert degree Kelvin to degree Celsius}
\usage{fConvertKtoC(Kelvin.V.n)}
\arguments{
  \item{Kelvin.V.n}{Data vector in Kelvin (degK)}
}

\value{Data vector in temperature Celsius (Temp_C, degC)}

\author{AMM
Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





