\name{fConvertKtoC}
\alias{fConvertKtoC}
\title{fConvertKtoC}
\description{Convert degree Kelvin to degree Celsius}
\usage{fConvertKtoC(Kelvin.V.n)}\arguments{
  \item{Kelvin.V.n}{Data vector in Kelvin (degK)}
}

\value{Data vector in temperature Celsius (Temp_C, degC)}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





