\name{fConvertVisibleWm2toPhotons}
\alias{fConvertVisibleWm2toPhotons}
\title{fConvertVisibleWm2toPhotons}
\description{Convert units of visible radiation from irradiance to photons flux}
\usage{fConvertVisibleWm2toPhotons(Wm2.V.n)}
\arguments{
  \item{Wm2.V.n}{Data vector in units of irradiance (W m-2)}
}

\value{Data vector in units of photons flux (PPFD, umol photons m-2 s-1)}

\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





