\name{fConvertVisibleWm2toPhotons}
\alias{fConvertVisibleWm2toPhotons}
\title{fConvertVisibleWm2toPhotons}
\description{Convert units of visible radiation from irradiance to photons flux}
\usage{fConvertVisibleWm2toPhotons(Wm2 = Wm2.V.n, 
    Wm2.V.n)}
\arguments{
  \item{Wm2}{Data vector in units of irradiance (W m-2)}
  \item{Wm2.V.n}{deprecated}
}

\value{Data vector in units of photons flux (PPFD, umol photons m-2 s-1)}

\author{Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





