\name{fillNAForward}
\alias{fillNAForward}
\title{fillNAForward}
\description{replace NA by value of previous record}
\usage{fillNAForward(x, firstValue = median(x, na.rm = FALSE))}
\arguments{
  \item{x}{numeric vector to fill NAs}
  \item{firstValue}{value to be used for NA at the beginning of x}
}



\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





