\name{fLloydTaylor}
\alias{fLloydTaylor}
\title{Temperature dependence of soil respiration after Equation 11 in Lloyd & Taylor (1994)}

\usage{fLloydTaylor(Tsoil.V.n, R_ref.n, T_ref.n, E_0.n)}
\arguments{
  \item{Tsoil.V.n}{Data vector of soil temperature in Kelvin (°K)}
  \item{R_ref.n}{Respiration rate at reference temperature}
  \item{T_ref.n}{Reference temperature in Kelvin (°K)}
  \item{E_0.n}{Activation temperature in Kelvin (°K)}
}

\value{Data vector of respiration rate}

\author{AMM
reference<<
Lloyd J, Taylor JA (1994) On the temperature dependence of soil respiration. Functional Ecology, 8, 315-323.
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





