% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileHandlingFormats.R
\name{fLoadAmeriflux22}
\alias{fLoadAmeriflux22}
\title{Read basic variables from Ameriflux standard (as of 2022) files}
\usage{
fLoadAmeriflux22(file_path, ...)
}
\arguments{
\item{file_path}{scalar string: the path to the csv file}

\item{...}{further arguments to \code{\link{read_csv}}}
}
\description{
Reads Variables from file into data.frame from file and passes
it to \code{\link{read_from_ameriflux22}}.
}
\seealso{
\code{\link{read_from_ameriflux22}} \code{\link{help_export}}
}
