\name{fSetFile}
\alias{fSetFile}
\title{fSetFile}
\description{Set file name with path and check if directory and/or file exists}
\usage{fSetFile(FileName.s, Dir.s, IO.b, CallFunction.s = "")}
\arguments{
  \item{FileName.s}{File name as a string}
  \item{Dir.s}{Directory as a string}
  \item{IO.b}{Input/output flag, TRUE for input, FALSE for output}
  \item{CallFunction.s}{Name (string) of the caller function for warnings}
}

\value{Returns name of file with complete path.}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





