\name{getFilledExampleDETha98Data}
\alias{getFilledExampleDETha98Data}
\title{getFilledExampleDETha98Data}
\description{Get or create the gapfilled version of the Example_DETha98 example data}
\usage{getFilledExampleDETha98Data(exampleDir = getREddyProcExampleDir())}
\arguments{
  \item{exampleDir}{the directory where the
cached filled example data is stored}
}

\value{example data.frame Example_DETha98 processed by gapfilling.}

\author{Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut Šigut <sigut.l@czechglobe.cz> [ctb]}





