\name{getFilledExampleDETha98Data}
\alias{getFilledExampleDETha98Data}
\title{getFilledExampleDETha98Data}
\description{Get or create the gapfilled version of the Example_DETha98 example data}
\usage{getFilledExampleDETha98Data(exampleDir = getREddyProcExampleDir())}
\arguments{
  \item{exampleDir}{the directory where the
cached filled example data is stored}
}

\value{example data.frame Example_DETha98 processed by gapfilling.}

\author{Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





