\name{gfComputeQualityFlags}
\alias{gfComputeQualityFlags}
\title{gfComputeQualityFlags}

\usage{gfComputeQualityFlags(winWidthDays, nCov)}
\arguments{
  \item{winWidthDays}{scalar integer: full windows length in days}
  \item{nCov}{scalar integer: number of covariates}
}

\value{integer vector with components #TODO}

\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





