\name{LightResponseCurveFitter}
\alias{LightResponseCurveFitter}
\title{LightResponseCurveFitter}

\description{
Constructs an instance of class \code{\link{LightResponseCurveFitter-class}} 
}
\usage{LightResponseCurveFitter(...)}
\arguments{
  \item{\dots}{
}
}



\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





