\name{LightResponseCurveFitter_getParameterInitials}
\alias{LightResponseCurveFitter_getParameterInitials}
\title{LightResponseCurveFitter getParameterInitials}
\description{return the prior distribution of parameters}
\usage{LightResponseCurveFitter_getParameterInitials(thetaPrior)}
\arguments{
  \item{thetaPrior}{numeric vector prior estimate of parameters}
}

\value{a numeric matrix (3, nPar) of initial values for fitting parameters}

\author{Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut Šigut <sigut.l@czechglobe.cz> [ctb]}





