\name{LightResponseCurveFitter_getParameterInitials}
\alias{LightResponseCurveFitter_getParameterInitials}
\title{LightResponseCurveFitter getParameterInitials}
\description{return the prior distribution of parameters}
\usage{LightResponseCurveFitter_getParameterInitials(thetaPrior)}
\arguments{
  \item{thetaPrior}{numeric vector prior estimate of parameters}
}

\value{a numeric matrix (3,nPar) of initial values for fitting parameters}

\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





