\name{LightResponseCurveFitter_getParameterInitials}
\alias{LightResponseCurveFitter_getParameterInitials}
\title{LightResponseCurveFitter getParameterInitials}
\description{return the prior distribution of parameters}
\usage{LightResponseCurveFitter_getParameterInitials(thetaPrior)}
\arguments{
  \item{thetaPrior}{numeric vector prior estimate of parameters}
}

\value{a numeric matrix (3, nPar) of initial values for fitting parameters}

\author{Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]}





