% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_nonrectangular.R
\name{NonrectangularLRCFitter_computeGPPGradient}
\alias{NonrectangularLRCFitter_computeGPPGradient}
\title{Gradient of [NonrectangularLRCFitter_predictGPP()]}
\usage{
NonrectangularLRCFitter_computeGPPGradient(Rg, Amax, alpha, logitconv)
}
\arguments{
\item{Rg}{photosynthetic flux density [umol / m2 / s] or Global Radiation}

\item{Amax}{[numeric] saturation (beta parameter) adjusted for effect of VPD}

\item{alpha}{[numeric] slope at Rg = 0}

\item{logitconv}{[numeric] -> logit-transformed convexity parameter}
}
\description{
Gradient of [NonrectangularLRCFitter_predictGPP()]
}
