% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_nonrectangular.R
\name{NonrectangularLRCFitter_computeLRCGradient}
\alias{NonrectangularLRCFitter_computeLRCGradient}
\title{Gradient of [NonrectangularLRCFitter_predictLRC()]}
\usage{
NonrectangularLRCFitter_computeLRCGradient(
  theta,
  Rg,
  VPD,
  Temp,
  VPD0,
  fixVPD = (k == 0),
  TRef = 15
)
}
\arguments{
\item{theta}{[numeric] -> parameter vector (
theta[1] = kVPD (k),
theta[2] = beta0 (beta),
theta[3] = alpha,
theta[4] = RRef (rb), # theta[4] = E0,
theta[5] = logitconv)}

\item{Rg}{photosynthetic flux density [umol / m2 / s] or Global Radiation}

\item{VPD}{Vapor Pressure Deficit [hPa]}

\item{VPD0}{[hPa] -> Parameters VPD0 fixed to 10 hPa according to
Lasslop et al 2010}

\item{fixVPD}{boolean scalar or vector of nrow(theta):
fixVPD if TRUE the VPD effect is not considered and VPD is not part
of the computation}

\item{TRef}{numeric scalar of Temperature (degree Celsius) for reference 
respiration `RRef`}

\item{Temp}{Temperature [degC]}
}
\description{
Gradient of [NonrectangularLRCFitter_predictLRC()]
}
\details{
This function computes the gradient of the Nonrectangular Light Response
Curve function with respect to the parameters.

Differs from base by extracting \code{conv} parameter from \code{theta} and adding
gradient to \code{logitconv} (3rd parameter from computeGPPGradient).
}
