\name{NonrectangularLRCFitter_getOptimizedParameterPositions}
\alias{NonrectangularLRCFitter_getOptimizedParameterPositions}
\title{NonrectangularLRCFitter getOptimizedParameterPositions}

\usage{NonrectangularLRCFitter_getOptimizedParameterPositions(isUsingFixedVPD, 
    isUsingFixedAlpha)}
\arguments{
  \item{isUsingFixedVPD}{
}
  \item{isUsingFixedAlpha}{
}
}



\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





