\name{NonrectangularLRCFitter_getOptimizedParameterPositions}
\alias{NonrectangularLRCFitter_getOptimizedParameterPositions}
\title{NonrectangularLRCFitter getOptimizedParameterPositions}
\description{get the positions of the parameters to optimize for given conditions on fixing alpha or VPD effect }
\usage{NonrectangularLRCFitter_getOptimizedParameterPositions(isUsingFixedVPD, 
    isUsingFixedAlpha)}
\arguments{
  \item{isUsingFixedVPD}{boolean scalar: if TRUE, VPD effect set to zero and is not optimized}
  \item{isUsingFixedAlpha}{boolean scalar: if TRUE, initial slope is fixed and is not optimized}
}



\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





