% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitioningLasslop10Nighttime.R
\name{partGLEstimateTempSensInBoundsE0Only}
\alias{partGLEstimateTempSensInBoundsE0Only}
\title{Estimate bounded temperature sensitivity E0 and RRef of ecosystem respiration}
\usage{
partGLEstimateTempSensInBoundsE0Only(
  REco,
  temperatureKelvin,
  prevE0 = NA,
  TRefFit = median(temperatureKelvin, na.rm = TRUE)
)
}
\arguments{
\item{REco}{night time NEE}

\item{temperatureKelvin}{temperature in Kelvin}

\item{prevE0}{numeric scalar: the previous guess of Temperature Sensitivity}

\item{TRefFit}{numeric scalar reference temperature for which RRef is estimated.
Set it to the center of the data (the default) for best fit (lowest uncertainty)}
}
\value{
list with entries
\itemize{
\item \code{E0}      : estimated temperature sensitivty E0 bounded to \link{50, 400}
\item \code{sdE0}    : standard deviation of E0
\item \code{TRefFit} : reference temperature used in the E0 fit
\item \code{RRefFit} : respiration at TRefFit
}
}
\description{
Estimate bounded temperature sensitivity E0 and RRef of ecosystem respiration
}
\seealso{
\code{\link[=partGLFitLRCWindows]{partGLFitLRCWindows()}}
}
