\name{partGLSmoothTempSens}
\alias{partGLSmoothTempSens}
\title{partGLSmoothTempSens}
\description{Smoothes time development of E0}
\usage{partGLSmoothTempSens(E0Win)}
\arguments{
  \item{E0Win}{data.frame with columns E0 and sdE0, RRefFit, and TRefFit with one row for each window}
}

\value{dataframe E0Win with updated columns E0 and sdE0}

\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





