% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitioningLasslop10.R
\name{partitionNEEGL}
\alias{partitionNEEGL}
\title{Partition NEE fluxes into GPP and Reco using the daytime method.}
\usage{
partitionNEEGL(
  ds,
  NEEVar = if (!missing(NEEVar.s)) NEEVar.s else paste0("NEE", suffixDash, "_f"),
  TempVar = if (!missing(TempVar.s)) TempVar.s else "Tair_f",
  VPDVar = if (!missing(VPDVar.s)) VPDVar.s else "VPD_f",
  RadVar = if (!missing(RadVar.s)) RadVar.s else "Rg_f",
  suffix = if (!missing(Suffix.s)) Suffix.s else "",
  NEEVar.s,
  TempVar.s,
  VPDVar.s,
  RadVar.s,
  Suffix.s,
  ...,
  controlGLPart = partGLControl(),
  isVerbose = TRUE,
  nRecInDay = 48L,
  lrcFitter = RectangularLRCFitter()
)
}
\arguments{
\item{ds}{dataset with all the specified input columns and full days in equidistant times}

\item{NEEVar}{Variable of NEE}

\item{TempVar}{Filled air or soil temperature variable (degC)}

\item{VPDVar}{Filled Vapor Pressure Deficit - VPD - (hPa)}

\item{RadVar}{Filled radiation variable}

\item{suffix}{string inserted into column names before identifier for NEE
column defaults}

\item{...}{further arguments to \code{\link[=partGLExtractStandardData]{partGLExtractStandardData()}}, such as \code{PotRadVar}}

\item{controlGLPart}{further default parameters, see \code{\link[=partGLControl]{partGLControl()}}}

\item{isVerbose}{set to FALSE to suppress output messages}

\item{nRecInDay}{number of records within one day (for half-hourly data its 48)}

\item{lrcFitter}{R5 class instance responsible for fitting the light response curve.
Current possibilities are \code{RectangularLRCFitter()}, \code{NonrectangularLRCFitter()},
and \code{LogisticSigmoidLRCFitter()}.}
}
\value{
A data.frame with columns:
\itemize{
\item \verb{Reco_DT_<suffix>}: predicted ecosystem respiration: mumol CO2/m2/s
\item \verb{GPP_DT_<suffix>}: predicted gross primary production mumol CO2/m2/s
\item \verb{<LRC>}: Further light response curve (LRC) parameters and their standard
deviation depend on the used LRC (e.g. for the non-rectangular LRC see
\code{NonrectangularLRCFitter_getParamNames}). They are estimated for windows and
are reported with the first record of the window.
\item \code{FP_VARnight}   : NEE filtered for nighttime records (others NA)
\item \code{FP_VARday}     : NEE filtered for daytime records (others NA)
\item \code{NEW_FP_Temp}   : temperature after filtering for quality flag, degree Celsius
\item \code{NEW_FP_VPD}    : vapour pressure deficit after filtering for quality flag, hPa
\item \code{FP_RRef_Night} : basal respiration estimated from nighttime (W / m2)
\item \code{FP_qc}         : quality flag:
\itemize{
\item 0 : good parameter fit,
\item 1 : some parameters out of range, required refit,
\item 2 : next parameter estimate is more than two weeks away
}
\item \code{FP_dRecPar}    : records until or after closest record that has a parameter estimate associated
\item \code{FP_errorcode}  : information why LRC-fit was not successful or was rejected, see result of \code{\link[=LRC_fitLRC]{LRC_fitLRC()}}
\item \code{FP_GPP2000}    : predicted GPP at VPD = 0 and PAR = 2000 : a surrogate for maximum photosynthetic capacity
\item \code{FP_OPT_VPD}    : list object of fitting results including \code{iOpt} and \code{covParms}
\item \code{FP_OPT_NoVPD}  : list object of fitting results with fit neglecting the VPD effect
}
}
\description{
Partition NEE fluxes into GPP and Reco using the daytime method.
}
\details{
The daytime-based partitioning of measured net ecosystem fluxes into
gross primary production (GPP) and ecosystem respiration (Reco).

The fit to the light-response-curve is done by default using the Rectangular
hyperbolic function, as in Lasslop et al. (2010).

Alternative fittings can be used by providing the corresponding subclass of
\code{LightResponseCurveFitter-class} to \code{lrcFitter} argument.

see \code{LRC_predictGPP()}
}
\references{
\enumerate{
\item Lasslop G, Reichstein M, Papale D, et al. (2010) Separation of net
ecosystem exchange into assimilation and respiration using a light
response curve approach: critical issues and global evaluation. Global
Change Biology, Volume 16, Issue 1, Pages 187-208
}
}
\seealso{
\code{\link[=partGL_FitNight_E0_RRef]{partGL_FitNight_E0_RRef()}}
}
