\name{RectangularLRCFitter-class}
\Rdversion{1.1}
\docType{class}
\alias{RectangularLRCFitter-class}
% \alias{RectangularLRCFitter}

\title{Class \code{"RectangularLRCFitter"}}
\description{
Common rectangular hyperbolic light-response curve fitting.
}
\section{Extends}{
Class \code{"\linkS4class{LightResponseCurveFitter}"}, directly.

All reference classes extend and inherit methods from \code{"\linkS4class{envRefClass}"}.

}
\author{TW}

\keyword{classes}
\section{Methods}{
  \describe{
    \item{\code{computeGPPGradient(Rg, Amax, alpha)}:}{ ~~ }
    \item{\code{predictGPP(Rg, Amax, alpha)}:}{ ~~ }
  }

The following methods are inherited (from the corresponding class):
predictGPP ("LightResponseCurveFitter"), getParameterNames ("LightResponseCurveFitter"), fitLRC ("LightResponseCurveFitter"), getPriorLocation ("LightResponseCurveFitter"), getPriorScale ("LightResponseCurveFitter"), getParameterInitials ("LightResponseCurveFitter"), optimLRCBounds ("LightResponseCurveFitter"), getOptimizedParameterPositions ("LightResponseCurveFitter"), optimLRCOnAdjustedPrior ("LightResponseCurveFitter"), isParameterInBounds ("LightResponseCurveFitter"), optimLRC ("LightResponseCurveFitter"), computeCost ("LightResponseCurveFitter"), predictLRC ("LightResponseCurveFitter"), computeLRCGradient ("LightResponseCurveFitter")
}
