\name{renameVariablesInDataframe}
\alias{renameVariablesInDataframe}
\title{renameVariablesInDataframe}
\description{Rename the column names of a data.frame according to a given mapping}
\usage{renameVariablesInDataframe(data.F, mapping = getBGC05ToAmerifluxVariableNameMapping())}
\arguments{
  \item{data.F}{data.frame whose columns should be renamped}
  \item{mapping}{named character vector:
specifying a renaming (name -> value)
of the variables, see e.g.
\code{\link{getAmerifluxToBGC05VariableNameMapping}}}
}



\author{TW,
Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut Šigut <sigut.l@czechglobe.cz> [ctb]}





