\name{renameVariablesInDataframe}
\alias{renameVariablesInDataframe}
\title{renameVariablesInDataframe}
\description{Rename the column names of a data.frame according to a given mapping}
\usage{renameVariablesInDataframe(data.F, mapping = getBGC05ToAmerifluxVariableNameMapping())}
\arguments{
  \item{data.F}{data.frame whose columns should be renamed}
  \item{mapping}{named character vector:
specifying a renaming (name -> value)
of the variables, see e.g.
\code{\link{getAmerifluxToBGC05VariableNameMapping}}}
}



\author{TW,
Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





