\name{RHLightResponseCostC}
\alias{RHLightResponseCostC}
\title{RHLightResponseCostC}

\usage{RHLightResponseCostC(theta, flux, sdFlux, parameterPrior, 
    sdParameterPrior, Rg, VPD, Temp, VPD0, fixVPD)}
\arguments{
  \item{theta}{
}
  \item{flux}{
}
  \item{sdFlux}{
}
  \item{parameterPrior}{
}
  \item{sdParameterPrior}{
}
  \item{Rg}{
}
  \item{VPD}{
}
  \item{Temp}{
}
  \item{VPD0}{
}
  \item{fixVPD}{
}
}



\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





