\name{RHLightResponseCostC}
\alias{RHLightResponseCostC}
\alias{_REddyProc_RHLightResponseCostC}
\title{RHLightResponseCostC}

\usage{RHLightResponseCostC(theta, flux, sdFlux, 
    parameterPrior, sdParameterPrior, Rg, 
    VPD, Temp, VPD0, fixVPD)}
    
\description{
	Computing residual sum of sqares for predictions vs. data of NEE implemented in C
}
    
\arguments{
  \item{theta}{
}
  \item{flux}{
}
  \item{sdFlux}{
}
  \item{parameterPrior}{
}
  \item{sdParameterPrior}{
}
  \item{Rg}{
}
  \item{VPD}{
}
  \item{Temp}{
}
  \item{VPD0}{
}
  \item{fixVPD}{
}
}



\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





