\name{sEddyProc_sCalcPotRadiation}
\alias{sEddyProc_sCalcPotRadiation}
\title{sEddyProc sCalcPotRadiation}
\description{compute potential radiation from position and time}
\usage{sEddyProc_sCalcPotRadiation(useSolartime = TRUE, 
    useSolartime.b)}
\arguments{
  \item{useSolartime}{
}
  \item{useSolartime.b}{by default corrects hour (given in local winter time)}
}

\value{column PotRad_NEW in sTEMP}

\author{Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav <U+0160>igut <sigut.l@czechglobe.cz> [ctb]}





