\name{sEddyProc_sExportData}
\alias{sEddyProc_sExportData}
\title{sEddyProc$sExportData - Export internal sDATA data frame}
\description{Export class internal sDATA data frame}
\usage{sEddyProc_sExportData()}

\value{Return data frame sDATA with time stamp shifted back to original.}

\author{AMM
Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut Šigut <sigut.l@czechglobe.cz> [ctb]}





