\name{sEddyProc_sExportResults}
\alias{sEddyProc_sExportResults}
\title{sEddyProc sExportResults}
\description{Export class internal sTEMP data frame with result columns}
\usage{sEddyProc_sExportResults(isListColumnsExported = FALSE)}
\arguments{
  \item{isListColumnsExported}{if TRUE export list columns in addition
to numeric columns,
such as the covariance matrices of the the day-time-partitioning LRC fits}
}

\value{Return data frame sTEMP with results.}

\author{AMM
Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





