\name{sEddyProc_sPlotDiurnalCycle}
\alias{sEddyProc_sPlotDiurnalCycle}
\title{sEddyProc$sPlotDiurnalCycle - Image with diurnal cycles of each month}
\description{Generates image in specified format ('pdf' or 'png') with diurnal cycles.}
\usage{sEddyProc_sPlotDiurnalCycle(Var.s, QFVar.s = "none", 
    QFValue.n = NA_real_, Format.s = "pdf", 
    Dir.s = "plots")}
\arguments{
  \item{Var.s}{Variable to plot}
  \item{QFVar.s}{Quality flag of variable to be filled}
  \item{QFValue.n}{Value of quality flag for data to plot}
  \item{Format.s}{Graphics file format ('pdf' or 'png')}
  \item{Dir.s}{Directory for plotting}
}



\author{KS, AMM
Department for Biogeochemical Integration at MPI-BGC, Jena, Germany <REddyProc-help@bgc-jena.mpg.de> [cph], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





