\name{sEddyProc_sSetLocationInfo}
\alias{sEddyProc_sSetLocationInfo}
\title{sEddyProc sSetLocationInfo}
\description{set Location and time Zone information to sLOCATION}
\usage{sEddyProc_sSetLocationInfo(Lat_deg.n, Long_deg.n, 
    TimeZone_h.n)}
\arguments{
  \item{Lat_deg.n}{Latitude in (decimal) degrees (-90 to +90)}
  \item{Long_deg.n}{Longitude in (decimal) degrees (-180 to +180)}
  \item{TimeZone_h.n}{Time zone (in hours) shift to UTC, e.g. +1 for Berlin}
}



\author{TW
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





