% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aEddy.R
\name{sEddyProc_update_ustarthreshold_columns}
\alias{sEddyProc_update_ustarthreshold_columns}
\title{Add columns reporting the uStar threshold for each scenario to sDATA}
\usage{
sEddyProc_update_ustarthreshold_columns()
}
\value{
side effect in .self$sDATA new columns Ustar_Thresh_<ustarsuffix>
}
\description{
Add columns reporting the uStar threshold for each scenario to sDATA
}
\seealso{
\code{\link{sEddyProc_sGetUstarScenarios}}
}
