\name{sExportData}
\alias{sExportData}
\title{sEddyProc$sExportData - Export internal sDATA data frame}
\description{Export class internal sDATA data frame}
\usage{sExportData()}

\value{Return data frame sDATA with time stamp shifted back to original.}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





