\name{sExportResults}
\alias{sExportResults}
\title{sEddyProc$sExportData - Export internal sTEMP data frame with result columns}
\description{Export class internal sTEMP data frame with result columns}
\usage{sExportResults(isListColumnsExported = FALSE)}
\arguments{
  \item{isListColumnsExported}{if TRUE export list columns in addition to numeric columns,
such as the covariance matrices of the the day-time-partitioning LRC fits}
}

\value{Return data frame sTEMP with results.}

\author{AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





