\name{sFillMDC}
\alias{sFillMDC}
\title{sEddyProc$sFillMDC - Gap filling with mean diurnal course}
\description{Mean diurnal course (MDC) algorithm based on average values within ± one hour of adjacent days}
\usage{sFillMDC(WinDays.i, Verbose.b = TRUE)}
\arguments{
  \item{WinDays.i}{Window size for filling in days}
  \item{Verbose.b}{Print status information to screen}
}

\value{MDC filling results in sTEMP data frame.}

\author{AMM 
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





