\name{sPlotDailySums}
\alias{sPlotDailySums}
\title{sEddyProc$sPlotDailySums - Image with daily sums of each year}
\description{Generates image in specified format ('pdf' or 'png') with daily sums, see also \code{\link{sPlotDailySumsY}}.}
\usage{sPlotDailySums(Var.s, VarUnc.s = "none", Format.s = "pdf", 
    Dir.s = "plots")}
\arguments{
  \item{Var.s}{(Filled) variable to plot}
  \item{VarUnc.s}{Uncertainty estimates for variable}
  \item{Format.s}{Graphics file format ('pdf' or 'png')}
  \item{Dir.s}{Directory for plotting}
}



\author{KS, AMM  
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





