\name{sPlotFingerprintY}
\alias{sPlotFingerprintY}
\title{sEddyProc$sPlotFingerprintY - Plot fingerprint of specified year}
\description{The fingerprint for a single year is plotted to the current device, scaled to all data.}
\usage{sPlotFingerprintY(Var.s, QFvar.s, QFvalue.n, Year.i)}
\arguments{
  \item{Var.s}{Variable to plot}
  \item{QFvar.s}{Quality flag of variable to be filled}
  \item{QFvalue.n}{Value of quality flag for data to plot}
  \item{Year.i}{Year to plot}
}



\author{KS, AMM
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





