\name{sPrintData}
\alias{sPrintData}
\title{sEddyProc$sPrintData - Print internal sDATA data frame}
\description{Print class internal sDATA data frame}
\usage{
sPrintData()
}

\value{Print first 100 rows of data frame sDATA.}

\author{AMM
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





