\name{sxClosePlot}
\alias{sxClosePlot}
\title{sEddyProc - Internal function}
\description{Close screens and save graphics device to file.}
\usage{
sxClosePlot(PlotFile.s)
}
\arguments{
  \item{PlotFile.s}{Name of opened graphics device}
}



\author{KS, AMM
(Antje M. Moffat, Kerstin Sickel, Thomas Wutzler)}





