\name{sxSetTitle}
\alias{sxSetTitle}
\title{sEddyProc - Internal function}
\description{Set title of plot.}
\usage{sxSetTitle(Var.s, QFVar.s = "none", QFValue.n = NA_real_, 
    Name.s)}
\arguments{
  \item{Var.s}{Variable to plot}
  \item{QFVar.s}{Quality flag of variable to be filled}
  \item{QFValue.n}{Value of quality flag for data to plot}
  \item{Name.s}{Name of plot}
}

\value{String with plot title}

\author{KS, AMM
(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





