\name{usGetAnnualSeasonUStarMap}
\alias{usGetAnnualSeasonUStarMap}
\title{usGetAnnualSeasonUStarMap}
\description{extract mapping season -> uStar columns from Distribution result}
\usage{usGetAnnualSeasonUStarMap(uStarTh)}
\arguments{
  \item{uStarTh}{result of \code{\link{sEddyProc_sEstUstarThresholdDistribution}}
or \code{\link{sEddyProc_sEstUstarThreshold}}$uStarTh}
}



\author{TW
Markus Reichstein <mreichstein@bgc-jena.mpg.de> [aut], Thomas Wutzler <twutz@bgc-jena.mpg.de> [aut, cre], Antje Maria Moffat <antje.moffat@bgc.mpg.de> [aut, trl], Kerstin Sickel <ksickel@bgc-jena.mpg.de> [ctb, trl], Olaf Menzer <omenzer@bgc-jena.mpg.de> [ctb], Mirco Migliavacca <mmiglia@bgc-jena.mpg.de> [aut], Ladislav Šigut <sigut.l@czechglobe.cz> [ctb]}





