\name{whichValueGreaterEqualC}
\alias{whichValueGreaterEqualC}
\title{whichValueGreaterEqualC}
\description{searches a sorted integer vector for the next element that is >= a threshold in fast C-code }
\usage{whichValueGreaterEqualC(x, threshold, iStart)}
\arguments{
  \item{x}{
}
  \item{threshold}{
}
  \item{iStart}{
}
}



\author{(Department for Biogeochemical Integration at MPI-BGC, Jena, Germany)}





